% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{add_density}
\alias{add_density}
\alias{layer_dnorm}
\title{Add a density curve to a plot}
\usage{
add_density(fun = dnorm, col = .ggblue, lwd = 1.5, lty = 2, ...)

layer_dnorm(x, sd = 1, mean = 0, ...)
}
\arguments{
\item{fun}{passed to \code{ggplot2::stat_function}}

\item{col}{passed to \code{ggplot2::stat_function}}

\item{lwd}{passed to \code{ggplot2::stat_function}}

\item{lty}{passed to \code{ggplot2::stat_function}}

\item{...}{passed to \code{ggplot2::stat_function}}

\item{x}{a \code{ggplot} object to which the density line will be added}

\item{sd}{passed to \code{stats::dnorm}}

\item{mean}{passed to \code{stats::dnorm}}
}
\description{
For example, a plot a normal density over a histogram
of conditional weighted residuals.
}
