% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{as.list.pm_opts}
\alias{as.list.pm_opts}
\title{Coerce pm_opts object to a list}
\usage{
\method{as.list}{pm_opts}(x, ...)
}
\arguments{
\item{x}{the \code{pm_opts} object}

\item{...}{not used}
}
\description{
Coerce pm_opts object to a list
}
\examples{
x <- as.list(pm)

}
