% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{boxwork}
\alias{boxwork}
\title{Make boxplots}
\usage{
boxwork(
  df,
  x,
  y,
  xs = defcx(),
  ys = defy(),
  fill = opts$boxplot.fill,
  alpha = opts$boxplot.alpha,
  hline = NULL,
  title = NULL,
  shown = TRUE,
  points = NULL,
  outlier.shape = opts$boxplot.outlier.shape,
  ...
)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{xs}{see \code{\link{defcx}}}

\item{ys}{see \code{\link{defy}}}

\item{fill}{passed to \code{\link[ggplot2:geom_boxplot]{ggplot2::geom_boxplot()}}. Specifying \code{NULL} disables
relaying the value to \code{geom_boxplot} and is required for modifying the fill
of the returned plot with \code{\link[ggplot2:aes]{ggplot2::aes()}}.}

\item{alpha}{passed to \code{geom_boxplot}}

\item{hline}{used to draw horizontal reference line}

\item{title}{passed to \code{ggtitle}}

\item{shown}{if \code{TRUE} provide a numeric summary of each
box (see details)}

\item{points}{show points in back of transparent boxes; if \code{TRUE},
a default display of points is made on top of boxes; also may be passed as
a list of arguments to pass to \code{geom_point}; see details}

\item{outlier.shape}{passed to \code{geom_boxplot}}

\item{...}{arguments passed to \code{geom_boxplot}}
}
\description{
Make boxplots
}
\details{
Since this function creates a boxplot,
the \code{x} column must be character, factor
or logical and \code{y} column must
be numeric.

If \code{shown} is \code{TRUE}, a numeric summary of each
box is included
below each box.  In the summary, \code{n} is the number of
non-NA observations in the \code{y} column for that box and
\code{N} is the number of unique \code{ID} values for
that box.  An error will be generated if \code{ID} does
not exist in the plotting data frame when \code{shown} is
\code{TRUE}.  When \code{N} is equal to \code{n} in the
summary, only \code{n} is shown.

The summaries will not be correct if the plot is eventually faceted by
another variable in the data set.  In this case, either use
\code{shown=FALSE} or create the plot with \code{\link{split_plot}}.

When the user passes the \code{points} argument, \code{outlier.shape} is
automatically switched to \code{NA} so that outlier points are only plotted
once.  The \code{fill} argument is also set to \code{NA}, so that boxes become
transparent, showing the points.

When the user sets \code{points} to \code{TRUE}, grey points are shown
in back of transparent boxes and the points are jittered in the x-direction.
The user can customize the display of the points by passing a list of
arguments for \code{geom_point} (for example, change the color, transparency,
size, jitter amount, etc). The user can also pass \code{jitter_width} in
the \code{points} list to set the amount of jitter in the x-direction while
keeping jitter in the y-direction zero. Passing \code{jitter_width} will
override any other value passed under \code{position} in the \code{points}
list.
}
