% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{$<-.pm_opts}
\alias{$<-.pm_opts}
\title{Assign a pmplots option}
\usage{
\method{$}{pm_opts}(x, i) <- value
}
\arguments{
\item{x}{the \code{pm_opts} object}

\item{i}{the name of the option to set}

\item{value}{the value to set}
}
\description{
Assign a pmplots option
}
\examples{

pm$scatter.size <- 4

}
