% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_label}
\alias{col_label}
\title{Column // axis-label specification}
\usage{
col_label(x)
}
\arguments{
\item{x}{string encoding data column and axis title}
}
\description{
The col-label specification is a way to identify a column
in a data set that is to be used for plotting (the \code{col} piece) along
with a label to be used for the axis data (the \code{label} piece).
}
\details{
Typically, the \code{col} part is separated from the \code{label}
part with a double front-slash.  So, to identify the column containing
the weight covariate we might specify \code{WT//Weight (kg)}.
}
\examples{
col_label("CL // Clearance (L)")
col_label("CL//Clearance (L)")
col_label("CL $$ Clearance (L)")
col_label("CL !! Clearance (L)")
col_label("CL @@ Clearance (L)")

try(col_label("CL / Clearance (L)"))

}
