% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{cont_cat}
\alias{cont_cat}
\alias{eta_cat}
\title{Plot continuous varible against categorical variable}
\usage{
cont_cat(df, x, y, xs = defcx(), ys = defy(), ...)

eta_cat(df, x, y, hline = 0, ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character col//title for x-axis data; see \code{\link{col_label}}}

\item{y}{character col//title for y-axis data; see \code{\link{col_label}}}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{...}{other arguments passed to \code{\link{boxwork}}}

\item{hline}{passed to \code{\link{boxwork}}}
}
\description{
Plot continuous varible against categorical variable
}
\details{
Since this function creates a boxplot,
the \code{x} column must be character, factor
or logical and \code{y} column must
be numeric.
}
\examples{
df <- dplyr::filter(superset2(), EVID==0)

cont_cat(df, x="STUDYc//Study name", y="WT//Weight (kg)")

}
\seealso{
\code{\link{boxwork}}
}
