% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displays.R
\name{cont_cat_panel}
\alias{cont_cat_panel}
\alias{cont_cat_panel_list}
\alias{cont_cat_panel_flat}
\title{Create a display of continuous versus categorical covariates}
\usage{
cont_cat_panel(
  df,
  x,
  y,
  ncol = 2,
  tag_levels = NULL,
  byrow = FALSE,
  transpose = FALSE,
  points = NULL,
  ...
)

cont_cat_panel_list(df, x, y, transpose = FALSE, ...)

cont_cat_panel_flat(...)
}
\arguments{
\item{df}{a data frame to plot.}

\item{x}{character \verb{col//title} for the categorical covariates to
plot on x-axis; see \code{\link[=col_label]{col_label()}}.}

\item{y}{character \verb{col//title} for the continuous covariates to
plot on y-axis; see \code{\link[=col_label]{col_label()}}.}

\item{ncol}{passed to \code{\link[=pm_grid]{pm_grid()}}.}

\item{tag_levels}{passed to \code{\link[patchwork:plot_annotation]{patchwork::plot_annotation()}}.}

\item{byrow}{passed to \code{\link[=pm_grid]{pm_grid()}}.}

\item{transpose}{logical; if \code{TRUE}, output will be transposed to
group plots by the categorical covariates rather than the continuous
covariates.}

\item{points}{passed to \code{\link[=boxwork]{boxwork()}} when plotting categorical covariates.}

\item{...}{passed to \code{\link[=cont_cat_panel_list]{cont_cat_panel_list()}}.}
}
\value{
\code{cont_cat_panel()} returns a list of plots arranged in graphics as a
\code{patchwork} object using \code{\link[=pm_grid]{pm_grid()}}. \code{cont_cat_panel_list()} returns the
same plots, but unarranged as a named list of lists. \code{cont_cat_panel_flat()}
calls \code{\link[=cont_cat_panel_list]{cont_cat_panel_list()}} and flattens the result with
\code{\link[purrr:flatten]{purrr::flatten()}}, returning a flat \code{pm_display} list of plots that can be arranged with
\code{\link[patchwork:wrap_plots]{patchwork::wrap_plots()}}.

When \code{transpose} is \code{FALSE} (default), plots in a single graphic are grouped
by the continuous covariates (passed as \code{y}), and the names of the list
reflect those names (e.g., \code{WT}). When \code{transpose} is \code{TRUE}, the graphics
are grouped by the categorical covariates (passed as \code{x}) and the names of
the list reflect those names (e.g. \code{RF}). See \strong{Examples}.
}
\description{
Get a single graphic of continuous covariate boxplots split by
levels of different categorical covariates (\code{cont_cat_panel()}).
Alternatively, get the component plots to be arranged by the user
(\code{cont_cat_panel_list()})
}
\details{
Pass \code{ncol = NULL} or another non-numeric value to bypass arranging plots
coming from \code{cont_cat_panel()}.
}
\examples{
data <- pmplots_data_id()
cont <- c("WT//Weight (kg)", "ALB//Albumin (mg/dL)", "AGE//Age (years)")
cats <- c("RF//Renal function", "CPc//Child-Pugh")

cont_cat_panel(data, x = cats, y = cont, tag_levels = "A")

cont_cat_panel(data, cats, cont)
cont_cat_panel(data, cats, cont, transpose = TRUE)

l <- cont_cat_panel_list(data, cats, cont, transpose = TRUE)
names(l)
with(l$RF, WT/(ALB + AGE), tag_levels = "A")

}
\seealso{
\code{\link[=eta_covariate]{eta_covariate()}}, \code{\link[=eta_covariate_list]{eta_covariate_list()}}
}
