% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatt.R
\name{cont_cont}
\alias{cont_cont}
\alias{eta_cont}
\alias{cwres_cont}
\alias{wres_cont}
\alias{res_cont}
\title{Plot continuous variable versus continuous variable}
\usage{
cont_cont(df, x, y, xs = defx(), ys = defy(), ...)

eta_cont(df, x, y, ...)

cwres_cont(df, x, y = "CWRES//Conditional weighted residual", xs = defx(),
  ys = defy(), ...)

wres_cont(df, x, y = "WRES//Weighted residual", xs = defx(), ys = defy(),
  ...)

res_cont(df, x, y = "RES//Residual", xs = defx(), ys = defy(), ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character col//title for x-axis data; see \code{\link{col_label}}}

\item{y}{character col//title for y-axis data; ; see \code{\link{col_label}}}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{...}{passed to \code{\link{scatt}}  and \code{\link{layer_hs}}}
}
\description{
This function is primarily called by other functions.
}
\details{
Since this function creates a scatter plot,
both the \code{x} and \code{y} columns must
be numeric.
}
\examples{
df <- dplyr::filter(superset2(), EVID==0)

cont_cont(df, x="WT//Weight (kg)", y="HT//Height (cm)")

}
\seealso{
\code{\link{scatt}}
}
