% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{cont_hist}
\alias{cont_hist}
\title{Generate a histogram plot}
\usage{
cont_hist(df, x, xs = defx(), fill = "black", col = "white",
  alpha = 0.6, ...)
}
\arguments{
\item{x}{the x column}

\item{xs}{a list of information for the x axis}

\item{fill}{a character value passed to \code{geom_histogram}}

\item{col}{a character value passed to \code{geom_histogram}}

\item{alpha}{a numeric value passed to \code{geom_histogram}}

\item{...}{passed to \code{geom_histogram}}
}
\description{
Generate a histogram plot
}
