% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displays.R
\name{covariate_split}
\alias{covariate_split}
\alias{covariate_split_list}
\title{Create a display of continuous versus categorical covariates}
\usage{
covariate_split(
  df,
  x,
  y,
  ncol = 2,
  tag_levels = NULL,
  byrow = FALSE,
  transpose = FALSE,
  ...
)

covariate_split_list(df, x, y, transpose = FALSE, ...)
}
\arguments{
\item{df}{a data frame to plot.}

\item{x}{character \verb{col//title} for the categorical covariates to
plot on x-axis; see \link{col_label}.}

\item{y}{character \verb{col//title} for the continuous covariates to
plot on y-axis; see \link{col_label}.}

\item{ncol}{passed to \code{\link[=pm_grid]{pm_grid()}}.}

\item{tag_levels}{passed to \code{\link[patchwork:plot_annotation]{patchwork::plot_annotation()}}.}

\item{byrow}{passed to \code{\link[=pm_grid]{pm_grid()}}.}

\item{transpose}{logical; if \code{TRUE}, output will be transposed to
group plots by the categorical covariates rather than the continuous
covariates.}

\item{...}{additional arguments passed to \code{\link[=cont_cat]{cont_cat()}}.}
}
\description{
Get a single graphic of continous covariate boxpots split by
levels of different categorical covariates (\code{covariate_split()}).
Alternatively, get the component plots to be arranged by the user
(\code{covariate_split_list()})
}
\details{
Pass \code{ncol = NULL} or another non-numeric value to bypass arranging plots
coming from \code{covariate_split()}.
}
\examples{
data <- pmplots_data_id()
cont <- c("WT//Weight (kg)", "ALB//Albumin (mg/dL)", "AGE//Age (years)")
cats <- c("RF//Renal function", "CPc//Child-Pugh")

covariate_split(data, x = cats, y = cont, tag_levels = "A")

covariate_split(data, cats, cont)
covariate_split(data, cats, cont, transpose = TRUE)

covariate_split_list(data, cats, cont)
covariate_split_list(data, cats, cont, transpose = TRUE)

}
\seealso{
\code{\link[=eta_covariate]{eta_covariate()}}, \code{\link[=eta_covariate_list]{eta_covariate_list()}}
}
