% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{cwres_cat}
\alias{cwres_cat}
\title{Plot CWRES versus categorical variable}
\usage{
cwres_cat(df, x, y = "CWRES//Conditional weighted residuals", hline = 0,
  ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{hline}{where to draw horizontal refrence line}

\item{...}{passed to \code{\link{cont_cat}}}
}
\description{
Plot CWRES versus categorical variable
}
\examples{
df <- dplyr::filter(superset2(), EVID==0)
cwres_cat(df, x="STUDYc//Study name")

}
