% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_time.R
\name{cwres_time}
\alias{cwres_time}
\alias{cwres_tafd}
\alias{cwres_tad}
\alias{cwresi_time}
\alias{cwresi_tafd}
\alias{cwresi_tad}
\title{Plot conditional weighted residuals versus time}
\usage{
cwres_time(df, x = pm_axis_time(), y = pm_axis_cwres(), ...)

cwres_tafd(df, x = pm_axis_tafd(), y = pm_axis_cwres(), ...)

cwres_tad(df, x = pm_axis_tad(), y = pm_axis_cwres(), ...)

cwresi_time(df, x = pm_axis_time(), y = pm_axis_cwresi(), ...)

cwresi_tafd(df, x = pm_axis_tafd(), y = pm_axis_cwresi(), ...)

cwresi_tad(df, x = pm_axis_tad(), y = pm_axis_cwresi(), ...)
}
\arguments{
\item{df}{data set to plot}

\item{x}{x-axis data in \code{\link{col_label}} format}

\item{y}{y-axis data in \code{\link{col_label}} format}

\item{...}{passed to \code{\link{res_time}} and eventually to
\code{\link{y_time}}}
}
\description{
Plot conditional weighted residuals versus time
}
\details{
See \code{\link{res_time}} for other important implementation details.
}
\examples{
df <- pmplots_data_obs()

cwresi_time(df)

}
\seealso{
\code{\link{res_time}}, \code{\link{wres_time}},
\code{\link{npde_time}}
}
