% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{defcx}
\alias{defcx}
\title{Default setting for discrete x-axis scale}
\usage{
defcx(...)
}
\arguments{
\item{...}{arguments for \code{scale_x_discrete}}
}
\description{
Basically the formal arguments for \code{scale_x_discrete}.
}
\examples{
defcx()

}
