% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{defx}
\alias{defx}
\title{Default setting for x-axis scale}
\usage{
defx(...)
}
\arguments{
\item{...}{arguments for \code{scale_x_continuous}}
}
\description{
Basically the formal arguments for \code{scale_x_continuous}.
}
\examples{
defx(trans="log")

}
