% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{defy}
\alias{defy}
\title{Default setting for continuous y-axis scale}
\usage{
defy(...)
}
\arguments{
\item{...}{arguments for \code{scale_y_continuous}}
}
\description{
Basically the formal arguments for \code{scale_y_continuous}.
}
\examples{
defy(trans="log")

}
