% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_pred.R
\name{dv_pred}
\alias{dv_pred}
\alias{dv_ipred}
\alias{dv_preds}
\title{Plot DV versus predicted values}
\usage{
dv_pred(df, x = "PRED//Population predicted {xname}", y = "DV//{yname}",
  yname = "value", xname = yname, xs = list(), ys = list(),
  loglog = FALSE, scales = c("fixed", "free"), ...)

dv_ipred(df, x = pm_axis("ipred"), ...)

dv_preds(df, ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{yname}{used to form y-axis label}

\item{xname}{used to form x-axis label}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{loglog}{if \code{TRUE}, x- and y-axes will be log-transformed}

\item{scales}{if \code{TRUE}, then the x- and y- axes will be forced
to have the same limits}

\item{...}{passed to \code{\link{scatt}} and \code{\link{layer_as}}}

\item{prefix}{used internally}
}
\description{
Plot DV versus predicted values
}
\details{
Since this function creates a scatter plot,
both the \code{x} and \code{y} columns must
be numeric.

\code{dv_preds} returns a list of two plots, with
the result of \code{dv_pred} in the first position
and the result of \code{dv_ipred} in the
second position.  In this case, \code{...} are
passed to both functions.
}
\examples{
df <- dplyr::filter(pmplots_data(), EVID==0)

dv_pred(df)

dv_ipred(df, yname="MyDrug (ng/mL)")

dv_preds(df, yname = "MyDrug (ng/mL)")

}
