% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_pred_ipred.R
\name{dv_pred_ipred}
\alias{dv_pred_ipred}
\alias{dv_pred_ipred_impl}
\alias{do_dv_pred_ipred}
\title{Create plots showing DV, PRED, and IPRED by individual}
\usage{
dv_pred_ipred(
  data,
  id_per_plot = 9,
  facets = pm_col_id(),
  nrow = NULL,
  ncol = NULL,
  fun = NULL,
  ...,
  id_col = deprecated()
)

dv_pred_ipred_impl(
  data,
  x = pm_axis_time(),
  dv = pm_axis_dv(),
  pred = pm_axis_pred(),
  ipred = pm_axis_ipred(),
  facets = pm_col_id(),
  id_col = deprecated(),
  xbreaks = waiver(),
  xunit = pm_opts$time.unit,
  xlab = NULL,
  ylab = NULL,
  angle = NULL,
  strip.text = element_text(),
  font_size = deprecated(),
  margin = deprecated(),
  plot.margin = NULL,
  legend.position = "top",
  pred_lty = 2,
  ipred_lty = 1,
  pred_point = TRUE,
  ipred_point = TRUE,
  lwd = 0.5,
  size = pm_opts$scatter.size,
  dv_shape = 19,
  dv_line = FALSE,
  dv_lwd = 0.5,
  scales = "free",
  log_y = FALSE,
  use_theme = pm_theme(),
  dv_color = "black",
  ipred_color = "red2",
  pred_color = "blue2",
  ncol = NULL,
  nrow = NULL,
  axis.text.rel = NULL,
  fun = NULL
)

do_dv_pred_ipred(data, options = list())
}
\arguments{
\item{data}{the data frame to plot.}

\item{id_per_plot}{number of unique combinations of \code{facets} columns to
include on each page.}

\item{facets}{a character vector of column names to use for faceting the
plot, passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}; if passed in \code{\link[=col_label]{col_label()}} format,
then the column data is modified with a call to \code{\link[glue:glue]{glue::glue_data()}}, so that
the strip label is more informative; see examples.}

\item{nrow}{number of rows in the plot grid; passed to
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}.}

\item{ncol}{number of columns in the plot grid; passed to
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}.}

\item{fun}{a function accepting a gg object as argument and returning
an updated gg object; if supplied, this function is applied to each plot
in the output list.}

\item{...}{additional arguments passed to \code{\link[=dv_pred_ipred_impl]{dv_pred_ipred_impl()}}.}

\item{id_col}{deprecated; use \code{facets} argument instead.}

\item{x}{the time-axis column, in
\code{\link[=col_label]{col_label()}} format; the title portion is
used for the x-axis title along with \code{xunit}; see also the \code{xlab} argument.}

\item{dv}{the \code{DV} column, in \code{\link[=col_label]{col_label()}} format; the title portion is used
for the y-axis title; see also the \code{ylab} argument.}

\item{pred}{the name of the \code{PRED} column; \code{\link[=col_label]{col_label()}} format is allowed,
but the label portion is discarded.}

\item{ipred}{the name of the \code{IPRED} column; \code{\link[=col_label]{col_label()}} format is allowed,
but the label portion is discarded.}

\item{xbreaks}{x-axis (time) breaks; passed to
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}.}

\item{xunit}{used to form x-axis title only if \code{xlab} is not provided.}

\item{xlab}{x-axis title; if not \code{NULL}, passed to \code{\link[ggplot2:labs]{ggplot2::xlab()}}.}

\item{ylab}{y-axis title; if not \code{NULL}, passed to \code{\link[ggplot2:labs]{ggplot2::ylab()}}.}

\item{angle}{rotation angle for x-axis tick labels; passed to \code{\link[=rot_x]{rot_x()}}.}

\item{strip.text}{optionally, the result of \code{\link[ggplot2:element]{ggplot2::element_text()}} to
format the strip text (e.g. change the font size or padding).}

\item{font_size}{deprecated.}

\item{margin}{deprecated.}

\item{plot.margin}{for the plot; passed \code{\link[ggplot2:theme]{ggplot2::theme()}}.}

\item{legend.position}{passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}

\item{pred_lty}{\code{PRED} linetype; passed to \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}.}

\item{ipred_lty}{\code{IPRED} linetype; passed to \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}.}

\item{pred_point}{logical; should points be plotted for \code{PRED}?}

\item{ipred_point}{logical; should points be plotted for \code{IPRED}?}

\item{lwd}{line width for \code{PRED} and \code{IPRED}; passed to
\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}.}

\item{size}{size of shape for \code{DV}, \code{IPRED} and \code{PRED}; passed
to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{dv_shape}{shape for \code{DV}; passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{dv_line}{logical; if \code{TRUE} then a line is added to the plot connecting
\code{DV} points.}

\item{dv_lwd}{line width for \code{DV}; passed to \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} as
\code{lwd}.}

\item{scales}{passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}.}

\item{log_y}{logical; if \code{TRUE} then y-axis is shown in log-scale.}

\item{use_theme}{a theme to use for the plot.}

\item{dv_color}{color to use for \code{DV} points.}

\item{ipred_color}{color to use for \code{IPRED} line.}

\item{pred_color}{color to use for \code{PRED} line.}

\item{axis.text.rel}{relative text size for axis text; use this to
selectively decrease font size for axis tick labels.}

\item{options}{a named list of options to pass to \link{dv_pred_ipred}}
}
\description{
\code{DV} is plotted with a symbol while \code{PRED} and \code{IPRED} are plotted with lines
and symbols. The plot is faceted by unique individual identifier (like \code{ID}
or \code{USUBJID}) as well as other faceting variables.
}
\examples{
data <- pmplots_data_obs()

p <- dv_pred_ipred(
  data,
  ylab = "Concentration (ng/mL)",
  nrow = 3, ncol = 3
)

p <- dv_pred_ipred(
  data,
  facets = c("ID", "STUDYc//Study: {STUDYc}")
)

}
