% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_time.R
\name{dv_time}
\alias{dv_time}
\alias{dv_tafd}
\alias{dv_tad}
\title{Plot DV versus time}
\usage{
dv_time(
  df,
  x = pm_axis_time(),
  y = pm_axis_dv(),
  xunit = opts$time.unit,
  yname = "DV",
  group = pm_col_id(),
  xs = list(),
  ys = list(),
  log = FALSE,
  xby = NULL,
  ...
)

dv_tafd(df, x = pm_axis_tafd(), ...)

dv_tad(df, x = pm_axis_tad(), ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{xunit}{time units; used to form x-axis label}

\item{yname}{used to form y-axis label}

\item{group}{passed to \code{\link{scatt}}}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{log}{if \code{TRUE}, y-axis will be log-transformed}

\item{xby}{interval for x-axis breaks}

\item{...}{passed to \code{\link{scatt}}}
}
\value{
A single plot.
}
\description{
Plot DV versus time
}
\details{
Plots are generated using \code{\link{y_time}},
which then calls \code{\link{scatt}}.

Since this function creates a scatter plot,
both the \code{x} and \code{y} columns must
be numeric.

If the data set includes a \code{BLQ} column,
the values in the \code{y} column are
set to \code{NA} when \code{BLQ} is
not equal to \code{0}.
}
\examples{
df <- pmplots_data_obs()

dv_time(df, yname="MyDrug (ng/mL)")

dv_time(df, yname="MyDrug (ng/mL)", xunit="day")


# Add a rug

df <- pmplots_data()

dv_time(df) + geom_rug(data = ~ dplyr::filter(.x,BLQ > 0))

}
