% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_cont.R
\name{eta_cont}
\alias{eta_cont}
\title{Plot ETAs versus continuous variables}
\usage{
eta_cont(df, x, y, ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character col//title for x-axis data; see \code{\link{col_label}}}

\item{y}{character col//title for y-axis data; see \code{\link{col_label}}}

\item{...}{other arguments passed to \code{\link{cont_cont}}}
}
\description{
This is a vectorized function; see details.
}
\details{
This is a vectorized functions; if either x or y
are a character vector (in col//title format), a list
of plots is returned.  If both x and y are length 1,
then a single plot object (not a list) is returned.
}
\examples{
df <- pmplots_data_id()

eta_cont(df, x = "WT//Weight (kg)", y = "ETA1//ETA-CL")

}
\seealso{
\code{\link{eta_cat}}
}
