% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_cont.R
\name{eta_labs}
\alias{eta_labs}
\alias{eta_col_labs}
\title{Make ETA labels and col-labels}
\usage{
eta_labs(..., .prefix = "ETA-")

eta_col_labs(..., .eta_n = seq_along(labs), .prefix = "ETA-")
}
\arguments{
\item{...}{unquoted parameter names}

\item{.prefix}{used to generate ETA axis label}

\item{.eta_n}{integer ETA numbers for which to form labels}
}
\description{
Make ETA labels and col-labels
}
\examples{
eta_labs(CL, V2, KA)

}
