% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_pairs.R
\name{eta_pairs}
\alias{eta_pairs}
\title{Pairs plot for ETAs using ggpairs}
\usage{
eta_pairs(x, etas, bins = 15, alpha = 0.6, fill = "black", col = "grey",
  shk = list())
}
\arguments{
\item{x}{plotting data.frame}

\item{etas}{character col//label for pairs data; see \code{\link{col_label}}}

\item{bins}{passed to \code{geom_histogram}}

\item{alpha}{passed to \code{geom_histogram}}

\item{fill}{passed to \code{geom_histogram}}

\item{col}{passed to \code{geom_histogram}}

\item{shk}{not used}
}
\value{
The result from a ggpairs call.
}
\description{
Pairs plot for ETAs using ggpairs
}
\details{
This funciton requires the \code{GGally} package to be installed.
}
\examples{

df <- dplyr::filter(pmplots_data(), EVID==0)
id <- dplyr::distinct(df, ID, .keep_all = TRUE)

eta_pairs(id, c("ETA1//ETA-CL", "ETA2//ETA-VC", "ETA3//ETA-KA"))

}
