% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etapairs.R
\name{etapairs}
\alias{etapairs}
\title{Pairs plot for ETAs using ggpairs}
\usage{
etapairs(x, etas)
}
\arguments{
\item{x}{plotting data.frame}

\item{etas}{character col//label for pairs data; see \code{\link{col_label}}}
}
\value{
The result from a ggpairs call.
}
\description{
Pairs plot for ETAs using ggpairs
}
\details{
This funciton requires the \code{GGally} package to be installed.
}
\examples{
id <- superset2() \%>\% filter(EVID==0) \%>\% distinct(ID, .keep_all = TRUE)
etapairs(id, c("ETA1//ETA-CL", "ETA2//ETA-VC", "ETA3//ETA-KA"))

}
