% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{geom_3s}
\alias{geom_3s}
\alias{layer_3s}
\title{Additional reference lines}
\usage{
geom_3s(lwd = 1.35, lty = 1, col = "darkgrey", yintercept = c(-3, 3),
  ...)

layer_3s(x, lwd = 1.35, lty = 1, col = "darkgrey", yintercept = c(-3,
  3), ...)
}
\arguments{
\item{lwd}{passed to \code{geom_hline}}

\item{lty}{passed to \code{geom_hline}}

\item{col}{passed to \code{geom_hline}}

\item{yintercept}{passed to \code{geom_hline}}

\item{...}{passed to \code{geom_hline}}

\item{x}{ggplot object}
}
\description{
Additional reference lines
}
