% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{layer}
\alias{layer}
\alias{layer_hs}
\alias{layer_sh}
\alias{layer_as}
\alias{layer_sa}
\alias{layer_dots}
\alias{geom_3s}
\alias{layer_3s}
\title{Layer functions}
\usage{
layer_hs(x, smooth = gs(), hline = gh(), ...)

layer_sh(x, smooth = gs(), hline = gh(), ...)

layer_as(x, smooth = gs(), abline = ga(), ...)

layer_sa(x, smooth = gs(), abline = ga(), ...)

layer_dots(x, ...)

geom_3s(x, lwd = 1.35, lty = 3, ...)

layer_3s(x, lwd = 1.35, lty = 3, ...)
}
\arguments{
\item{smooth}{list of arguments for \code{geom_smooth}}

\item{hline}{list of arguments for \code{geom_hline}}

\item{abline}{list of arguments for \code{geom_abline}}
}
\description{
Layer functions
}
