% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_plot.R
\name{list_plot_x}
\alias{list_plot_x}
\alias{list_plot_y}
\alias{list_plot_xy}
\title{Apply a plotting function to a vector of x or y values}
\usage{
list_plot_x(df, x, y, .fun = cont_cont, ...)

list_plot_y(df, x, y, .fun = cont_cont, ...)

list_plot_xy(df, x, y, .fun = cont_cont, ...)
}
\arguments{
\item{df}{the plotting data set}

\item{x}{the x-column, as a col_label}

\item{y}{the y-column, as col_label}

\item{.fun}{the function to call}

\item{...}{passed to \code{fun}}
}
\value{
a list of plots generated from \code{.fun}
}
\description{
Apply a plotting function to a vector of x or y values
}
\details{
This function was intended for use with functions
that ultimately call \code{\link{cont_cont}} and
have continuous data on the x and y axes.
}
