% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_scale}
\alias{log_scale}
\title{Scale information for log transformation}
\usage{
log_scale(br = logbr(), ...)
}
\arguments{
\item{br}{breaks}

\item{...}{additional parameters}
}
\description{
Scale information for log transformation
}
\examples{
log_scale()

}
