% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrggsave.R
\name{mrggsave}
\alias{mrggsave}
\alias{mrggdraw}
\title{Save plot objects as .pdf file after labeling with Source graphic and Source code labels.}
\usage{
mrggsave(x, script, stem = "Rplot", dir = "../deliv/figure",
  prefix = gsub("^\\\\.\\\\./", "./", dir), onefile = TRUE,
  arrange = FALSE, labsep = "\\n", fontsize = 7, textGrob.x = 0.01,
  textGrob.y = 1, margin = c(0.1, 0.1, 0.7, 0.1), unit = "cm",
  draw = FALSE, nosave = FALSE, ...)

mrggdraw(...)
}
\arguments{
\item{x}{an object or list of objects of class \code{gg}}

\item{script}{the name of the script generating the \code{gg} objects}

\item{stem}{to form the name of the output \code{.pdf} file}

\item{dir}{output directory for \code{.pdf} file}

\item{prefix}{gets prepended to the output file path in the Source graphic label}

\item{onefile}{passed to \code{\link{pdf}}}

\item{arrange}{logical; if \code{TRUE}, arrange the ggplot objects on a single page with \code{arrangeGrob}}

\item{labsep}{character separator (or newline) for Source code and Source graphic labels}

\item{fontsize}{for Source graphic and Source code labels}

\item{textGrob.x}{passed to \code{textGrob} (as \code{x})}

\item{textGrob.y}{passed to \code{textGrob} (as \code{y})}

\item{margin}{numeric vector of length 4 or 1 to set top, right, bottom, left margin}

\item{unit}{unit to go along with margin sizes}

\item{draw}{if \code{TRUE}, the image is printed but not saved}

\item{nosave}{logical; if \code{TRUE}, return the labeled objects}

\item{...}{other arguments passed to \code{\link{pdf}} or \code{arrangeGrob}}
}
\description{
Save plot objects as .pdf file after labeling with Source graphic and Source code labels.
}
\details{
\itemize{
  \item This function will \code{\link{require}} the \code{gridExtra} package


}
}
\examples{

data(Theoph)
require(ggplot2)

Script <- "example.R"

## NOTE: see default value for dir argument, which should be appropriate for project work
## Changing it here only for the example
dir <- "."


p1 <- ggplot(data=Theoph) + geom_line(aes(x=Time, y=conc, group=Subject))
p2 <- ggplot(data=Theoph) + geom_line(aes(x=Time, y=conc)) + facet_wrap(~Subject)

mrggsave(p1, Script, "plot1", dir=dir)
mrggsave(p2, Script, "plot2", dir=dir)

mrggsave(list(p1,p2), Script, "both_plots", dir=dir)
mrggsave(list(p1,p2), Script, "separate_files", onefile=FALSE, dir=dir)

mrggsave(p1, Script, "different_shape", width=10, height=4, dir=dir)

mrggsave(list(p1,p2), Script, "onepage", dir=dir, arrange=TRUE, nrow=1, ncol=2)

}
