% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displays.R
\name{npde_covariate}
\alias{npde_covariate}
\alias{npde_covariate_list}
\title{Create NPDE versus covariate displays}
\usage{
npde_covariate(df, x, ncol = 2, tag_levels = NULL, byrow = NULL, points = NULL)

npde_covariate_list(df, x, points = NULL)
}
\arguments{
\item{df}{a data frame to plot.}

\item{x}{character \verb{col//title} for covariates to plot on x-axis;
see \code{\link[=col_label]{col_label()}}.}

\item{ncol}{passed to \code{\link[=pm_grid]{pm_grid()}}.}

\item{tag_levels}{passed to \code{\link[patchwork:plot_annotation]{patchwork::plot_annotation()}}.}

\item{byrow}{passed to \code{\link[=pm_grid]{pm_grid()}}.}

\item{points}{passed to \code{\link[=boxwork]{boxwork()}} when plotting categorical covariates.}
}
\value{
\code{npde_covariate()} returns single graphic of scatter plot diagnostics
as a \code{patchwork} object that has been arranged using \code{\link[=pm_grid]{pm_grid()}} and
\code{npde_covariate_list()} returns the same component plots unarranged in a list.
}
\description{
Get a single graphic of \code{NPDE} versus continuous and / or categorical
covariates (\code{npde_covariate()}) or get a list that can be arranged by the
user  (\code{npde_covariate_list()}). See \code{\link[=npde_panel]{npde_panel()}} for other \code{NPDE}
diagnostic displays.
}
\details{
Pass \code{ncol = NULL} or another non-numeric value to bypass arranging plots
coming from \code{npde_covariate()}.
}
\examples{
data <- pmplots_data_id()
cont <- c("WT//Weight (kg)", "ALB//Albumin (mg/dL)")
cats <- c("RF//Renal function", "CPc//Child-Pugh")

npde_covariate(data, x = c(cont, cats), tag_levels = "A")
npde_covariate_list(data, x = cats)


}
\seealso{
\code{\link[=cwres_covariate]{cwres_covariate()}}, \code{\link[=eta_covariate]{eta_covariate()}}
}
