% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{npde_ref}
\alias{npde_ref}
\title{Input parameters for NPDE reference lines}
\usage{
npde_ref(y = 0, lwd = 1, ...)
}
\arguments{
\item{y}{used to set yintercept}

\item{lwd}{width of reference lines}

\item{...}{other argument to set for \code{geom_hline}}
}
\description{
Input parameters for NPDE reference lines
}
