% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displays.R
\name{npde_scatter}
\alias{npde_scatter}
\alias{cwres_scatter}
\title{Create a display of residuals versus time and predicted values}
\usage{
npde_scatter(
  df,
  xname = "value",
  unit_time = opts$time.unit,
  unit_tad = opts$time.unit,
  xby_time = NULL,
  xby_tad = NULL,
  tag_levels = NULL,
  compact = FALSE
)

cwres_scatter(
  df,
  xname = "value",
  unit_time = opts$time.unit,
  unit_tad = opts$time.unit,
  xby_time = NULL,
  xby_tad = NULL,
  tag_levels = NULL,
  compact = FALSE
)
}
\arguments{
\item{df}{a data frame to plot.}

\item{xname}{passed to \code{\link[=npde_pred]{npde_pred()}}.}

\item{unit_time}{passed to \code{\link[=npde_tad]{npde_tad()}} as \code{xunit}.}

\item{unit_tad}{passed to \code{\link[=npde_time]{npde_time()}} as \code{xunit}.}

\item{xby_time}{passed to \code{\link[=npde_time]{npde_time()}} as \code{xby}.}

\item{xby_tad}{passed to \code{\link[=npde_tad]{npde_tad()}} as \code{xby}.}

\item{tag_levels}{passed to \code{\link[patchwork:plot_annotation]{patchwork::plot_annotation()}}.}

\item{compact}{use \code{compact = TRUE} to get a more compact display; see
\strong{Examples}.}
}
\value{
A single graphic with three panels (\code{NPDE} or \code{CWRES} versus \code{TIME}, \code{TAD}
and \code{PRED}) as a \code{patchwork} object. The default behavior is to create a
graphic with three panels in three rows, filling a portrait page.
Use \code{compact = TRUE} for a single graphic in two rows, with the \code{TIME}
plot on the top and the \code{TAD} and \code{PRED} plots on the bottom.
}
\description{
Get a single graphic of \code{NPDE} or \code{CWRES} diagnostics versus \code{TIME},
\code{TAD} and \code{PRED}. Output can be in either long or compact format.
}
\examples{
data <- pmplots_data_obs()
npde_scatter(data)
npde_scatter(data, compact = TRUE)

}
\seealso{
\code{\link[=npde_panel]{npde_panel()}}, \code{\link[=npde_panel_list]{npde_panel_list()}}, \code{\link[=cwres_panel]{cwres_panel()}},
\code{\link[=cwres_panel_list]{cwres_panel_list()}}
}
