% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_time.R
\name{npde_time}
\alias{npde_time}
\alias{npde_tad}
\alias{npde_tafd}
\title{Plot NPDE versus time}
\usage{
npde_time(df, x = pm_axis_time(), y = pm_axis_npde(), ...,
  hline = npde_ref())

npde_tad(df, x = pm_axis_tad(), y = pm_axis_npde(), ...)

npde_tafd(df, x = pm_axis_tafd(), y = pm_axis_npde(), ...)
}
\arguments{
\item{df}{data set to plot}

\item{x}{x-axis data in \code{\link{col_label}} format}

\item{y}{y-axis data in \code{\link{col_label}} format}

\item{...}{passed to \code{\link{res_time}} and eventually to
\code{\link{y_time}}}

\item{hline}{a list of arguments to pass to \code{geom_hline} specifying
aesthetics to use}
}
\description{
Uses \code{\link{y_time}} to create the plots.  Please see that
help topic and details here for other arguments to customize the plot.
}
\details{
Plots are generated using \code{\link{y_time}},
which then calls \code{\link{scatt}}.

By default, the time unit is assumed
to be hours (\code{hr}).  See the \code{xunit} argument
to \code{\link{y_time}} to change the time unit.

See the \code{xby} argument to \code{\link{y_time}} for a
convenient way to change the breaks for the x-axis (time).

Since this function creates a scatter plot,
both the \code{x} and \code{y} columns must
be numeric.
}
\examples{
df <- pmplots_data_obs()

npde_time(df)

}
\seealso{
\code{\link{res_time}}, \code{\link{cwres_time}},
\code{\link{wres_time}}
}
