% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_pairs.R
\name{pairs_lower_plot}
\alias{pairs_lower_plot}
\title{Default plots for lower diagonal}
\usage{
pairs_lower_plot(p)
}
\arguments{
\item{p}{a \code{gg} object}
}
\value{
A \code{gg} object, modified
}
\description{
Default plots for lower diagonal
}
\details{
\itemize{
\item the function should take one argument (a \code{gg} object`)
\item modify that object to render the data; this is typically using
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}
\item add aesthetics or layers as you would for any \code{ggplot} workflow
}
}
\examples{
data <- pmplots_data_id()
p <- ggplot(data = data, aes(WT, BMI))
p
pmplots:::pairs_lower_plot(p)

}
\keyword{internal}
