% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_label}
\alias{parse_label}
\alias{label_parse_label}
\title{Parse the label part of a col_label}
\usage{
parse_label(x)

label_parse_label(x)
}
\arguments{
\item{x}{a character string}
}
\description{
Parse the label part of a col_label
}
\examples{
parse_label("foo $\\\\mu$")

}
