% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{pm_aliases}
\alias{pm_aliases}
\alias{pm_set_aliases}
\alias{pm_clear_aliases}
\alias{pm_show_canonical}
\title{Set and manage aliases to certain canonical names used by pmplots}
\usage{
pm_aliases()

pm_set_aliases(...)

pm_clear_aliases()

pm_show_canonical()
}
\arguments{
\item{...}{\code{alias_name = canonical_name} pairs, where the right-hand side
can be quoted or unquoted.}
}
\description{
Some column names are assumed when calling pmplots plotting functions.
These names are generated by the functions listed in \link{pm_axis_functions}.
When these functions are invoked to select the data columns used for
plotting, hard-coded assumptions by pmplots are made (e.g., time is \code{TIME}).
When user-defined aliases are set, these aliases are used in place of the
canonical names assumed by pmplots. Aliases are only applied when these
\link{pm_axis_functions} are invoked to select the data column.
See \strong{Examples}.
\itemize{
\item \code{pm_aliases()} prints the currently active aliases.
\item \code{pm_set_aliases()} registers one or more aliases mapping a data column name
to a canonical pmplots column name.
\item \code{pm_clear_aliases()} removes all registered aliases.
\item \code{pm_show_canonical()} returns the canonical column names that can be aliased.
}
}
\examples{

# The canonical name for time is `TIME`
pm_axis_time()

# If you want to redirect to `TAFD` whenever the canonical name is `TIME`
# you can set an alias
pm_set_aliases(TAFD = TIME)

# Now, whenever canonical time is requested, pmplots redirects to `TAFD`
pm_axis_time()

# See current aliases
pm_aliases()

# Clear aliases
pm_clear_aliases()

# There is no canonical name for `WT`, so this will fail
try(pm_set_aliases(WEIGHT = WT))

# Because this call does not invoke `pm_axis_time`, no alias redirection
# will occur; in this case the user should compute the alias and provide 
# the requested column name (e.g. `TAFD`) in the call. 
data <- pmplots_data_obs()
dv_time(data, x = "TIME")

}
\seealso{
\link{pm_axis_functions}
}
