% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm-gg-labs.R
\name{pm_col_break}
\alias{pm_col_break}
\title{Make aesthetic labels break across two lines}
\usage{
pm_col_break(...)
}
\arguments{
\item{...}{\code{name = value} pairs; names are mapping whose label text will be
broken into two lines; values are the target lengths of the first lines
after breaking.}
}
\description{
Make aesthetic labels break across two lines
}
