% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm-gg-labs.R
\name{pm_label_columns}
\alias{pm_label_columns}
\title{Add axis label data to a data frame}
\usage{
pm_label_columns(df, spec, labs = list(), short_max = Inf)
}
\arguments{
\item{df}{a data frame.}

\item{spec}{a named list of label data; names correspond to columns
in the data used to make the plot; may also be a \code{yspec} object, which
will be converted to a named list through \code{\link[yspec:ys_get_short_unit]{yspec::ys_get_short_unit()}}.}

\item{labs}{another named list of label data to override names found in
\code{spec}.}

\item{short_max}{passed to \code{\link[yspec:ys_get_short_unit]{yspec::ys_get_short_unit()}}.}
}
\description{
Add axis label data to a data frame
}
