% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pm_log}
\alias{pm_log}
\alias{pm_ident}
\title{Identity and log scale helpers}
\usage{
pm_log(breaks = NULL, limits = NULL, trans = "log10", ...)

pm_ident(breaks, limits = range(breaks), ...)
}
\arguments{
\item{breaks}{passed to scale function}

\item{limits}{passed to scale function}

\item{trans}{passed to scale function}

\item{...}{passed to scale function}
}
\description{
Identity and log scale helpers
}
