% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont.R
\name{pm_scatter}
\alias{pm_scatter}
\alias{pm_scatter_list}
\alias{cont_cont}
\title{Plot continuous variable versus continuous variable}
\usage{
pm_scatter(df, x, y, xs = defx(), ys = defy(), ...)

pm_scatter_list(df, x, y, ...)

cont_cont(df, x, y, xs = defx(), ys = defy(), ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character col//title for x-axis data; see \code{\link{col_label}}}

\item{y}{character col//title for y-axis data; see \code{\link{col_label}}}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{...}{passed to \code{\link{scatt}}  and \code{\link{layer_hs}}}
}
\description{
This function is primarily called by other functions.
\code{pm_scatter} is an alias to \code{cont_cont} and should be
used in production code. \code{pm_scatter_list} is a vectorized
form of \code{pm_scatter}.
}
\details{
Since this function creates a scatter plot,
both the \code{x} and \code{y} columns must
be numeric.
}
\examples{
df <- pmplots_data_id()

pm_scatter(
  df,
  x="WT//Weight (kg)",
  y="HT//Height (cm)"
)

}
\seealso{
\code{\link{scatt}}
}
