% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_theme.R
\name{pm_smooths}
\alias{pm_smooths}
\title{Add loess line}
\usage{
pm_smooths(
  method = opts$smooth.method,
  se = FALSE,
  lty = opts$smooth.lty,
  lwd = opts$smooth.lwd,
  col = opts$smooth.col,
  ...
)
}
\arguments{
\item{method}{passed to \code{geom_smooth}}

\item{se}{passed to \code{geom_smooth}}

\item{lty}{passed to \code{geom_smooth}}

\item{lwd}{passed to \code{geom_smooth}}

\item{col}{passed to \code{geom_smooth}}

\item{...}{passed to \code{geom_smooth}}
}
\description{
Add loess line
}
