% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_theme.R
\name{pm_theme}
\alias{pm_theme}
\alias{pm_abline}
\alias{pm_smooth}
\alias{pm_hline}
\title{The standard pmplots theme}
\usage{
pm_theme()

pm_abline(...)

pm_smooth(...)

pm_hline(...)
}
\arguments{
\item{...}{passed to \code{\link{gs}} (smooth), \code{\link{gh}} (hline),
or \code{\link{ga}} (abline).}
}
\description{
The standard pmplots theme
}
\details{
\code{pm_theme} is an alias to \code{theme_bw}.
}
\examples{

data <- pmplots_data_obs()

ggplot(data,aes(PRED,DV)) + geom_point() +
  pm_theme() + pm_smooth() + pm_abline()

}
