% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pm_with}
\alias{pm_with}
\title{Arrange a named list of plots}
\usage{
pm_with(x, expr, tag_levels = NULL)
}
\arguments{
\item{x}{a named list of gg objects to arrange.}

\item{expr}{a \code{patchwork} formula for arranging plots in \code{x}.}

\item{tag_levels}{passed to \code{\link[patchwork:plot_annotation]{patchwork::plot_annotation()}}.}
}
\description{
Arrange a named list of plots
}
\examples{
data <- pmplots_data_id()
etas <- paste0("ETA", 1:3)
covs <- c("WT", "AGE", "ALB")

x <- eta_covariate_list(data, x = covs, y = etas, transpose = TRUE)

pm_with(x$WT, (ETA1 + ETA2) / ETA3)

}
