% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pmplots_data}
\alias{pmplots_data}
\alias{pmplots_data_id}
\alias{pmplots_data_obs}
\title{Example data sets}
\usage{
pmplots_data()

pmplots_data_id()

pmplots_data_obs()
}
\description{
Example data sets
}
\examples{

head(pmplots_data_obs())

head(pmplots_data_id())

}
