% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pmplots_nlmixr2_data}
\alias{pmplots_nlmixr2_data}
\title{Get the correct dataset for a nlmixr2 fit}
\usage{
pmplots_nlmixr2_data(data, type = c("left", "inner", "right"))
}
\arguments{
\item{data}{This is the dataset to check}

\item{type}{gives the type of merge between the original dataset
and the fit dataset.  The left dataset represents the original
dataset while the right dataset represents the fitted dataset.}
}
\value{
If this is a nlmixr2 fit object, return the expanded full
dataset.  If it isn't a nlmixr2 fit simply return the dataset
}
\description{
Get the correct dataset for a nlmixr2 fit
}
\examples{
if (requireNamespace("nlmixr2extra", quietly = TRUE)) {

 # This gets the merged dataset from a nlmixr2 fit
 # If we grab the pre-run nlmixr2 fit from nlmixr2extra we can see the fit doesn't have WT in it
 print(nlmixr2extra::theoFitOde)

 # But if we simply expand the dataset, the WT is added from the original dataset
 pmplots_nlmixr2_data(nlmixr2extra::theoFitOde)

}
}
\author{
Matthew Fidler
}
