% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{print.pm_opts}
\alias{print.pm_opts}
\title{Print the pm_opts object}
\usage{
\method{print}{pm_opts}(x, ...)
}
\arguments{
\item{x}{the \code{pm_opts} object}

\item{...}{not used}
}
\description{
Print the pm_opts object
}
\examples{
pm

}
