% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{res_cat}
\alias{res_cat}
\alias{wres_cat}
\alias{cwres_cat}
\title{Plot residuals versus categorical variable}
\usage{
res_cat(df, x, y = "RES//Residual", hline = 0, ...)

wres_cat(df, x, y = "WRES//Weighted residual", hline = 0, ...)

cwres_cat(df, x, y = "CWRES//Conditional weighted residual", hline = 0, ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{hline}{where to draw horizontal refrence line}

\item{...}{passed to \code{\link{cont_cat}}}
}
\description{
Plot residuals versus categorical variable
}
\details{
Since this function creates a boxplot,
the \code{x} column must be character, factor
or logical and \code{y} column must
be numeric.

Summary numbers located below each box are described in
\code{\link{boxwork}}.
}
\examples{
df <- dplyr::filter(superset2(), EVID==0)
cwres_cat(df, x="STUDYc//Study name")

}
\seealso{
\code{\link{boxwork}}
}
