% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_cont.R
\name{res_cont}
\alias{res_cont}
\alias{wres_cont}
\alias{cwres_cont}
\title{Plot residuals versus continuous variables}
\usage{
res_cont(df, x, y = "RES//Residual", xs = defx(), ys = defy(), ...)

wres_cont(df, x, y = "WRES//Weighted residual", xs = defx(), ys = defy(),
  ...)

cwres_cont(df, x, y = "CWRES//Conditional weighted residual", xs = defx(),
  ys = defy(), ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character col//title for x-axis data; see
\code{\link{col_label}}}

\item{y}{character col//title for y-axis data; see
\code{\link{col_label}}}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{...}{passed to \code{\link{scatt}}  and
\code{\link{layer_hs}}}
}
\description{
These are vectorized functions; see details;
}
\details{
These are vectorized functions; if either x or y
are a character vector (in col//title format), a list
of plots is returned.  If both x and y are length 1,
then a single plot object (not a list) is returned.
}
