% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_hist.R
\name{res_hist}
\alias{res_hist}
\alias{wres_hist}
\alias{cwres_hist}
\alias{cwresi_hist}
\alias{npde_hist}
\title{Histograms of residuals or NPDE}
\usage{
res_hist(df, ..., x = "RES", xname = "Residual", xs = defx())

wres_hist(df, ..., x = "WRES", xname = "Weighted residual")

cwres_hist(df, ..., x = "CWRES", xname = "Conditional weighted residual")

cwresi_hist(df, x = "CWRESI", ...)

npde_hist(df, ..., x = "NPDE",
  xname = "Normalized prediction distribution error")
}
\arguments{
\item{df}{data frame to plot}

\item{...}{passed to \code{\link{cont_hist}}}

\item{x}{character name for x-axis data}

\item{xname}{used to form x-axis label}

\item{xs}{see \code{\link{defx}}}
}
\description{
Histograms of residuals or NPDE
}
