% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefab.R
\name{res_hist_q}
\alias{res_hist_q}
\alias{npde_hist_q}
\alias{cwres_hist_q}
\title{Create a display of residual histograms and quantile-quantile plots}
\usage{
npde_hist_q(df, ncol = 1, tag_levels = NULL)

cwres_hist_q(df, ncol = 1, tag_levels = NULL)
}
\arguments{
\item{df}{a data frame to plot.}

\item{ncol}{passed to \code{\link[=pm_grid]{pm_grid()}}.}

\item{tag_levels}{passed to \code{\link[patchwork:plot_annotation]{patchwork::plot_annotation()}}.}
}
\value{
A single graphic is returned, with a \code{NPDE}  or \code{CWRES} histogram and
quantile-quantile plot arranged in a an object
using \code{\link[patchwork:plot_annotation]{patchwork::plot_annotation()}}.
}
\description{
Get a single graphic showing \code{NPDE} or \code{CWRES} histogram and
quantile-quantile plots.
}
\details{
The default value for \code{ncol} (1) means the two plots will be arranged in a
single column, with the histogram on the top and quantile-quantile plot
on the bottom. Using \code{ncol=2} will return an graphic with the plots
side by side.
}
\examples{
data <- pmplots_data_obs()
npde_hist_q(data, tag_levels = "a")
npde_hist_q(data, tag_levels = "a", ncol = 2)

}
