% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatt.R
\name{res_pred}
\alias{res_pred}
\alias{cwres_pred}
\alias{wres_pred}
\title{Residuals versus predicted values}
\usage{
res_pred(df, x = "PRED", y = "RES", xs = defx(), ys = defy(),
  xname = "value", ...)

cwres_pred(df, x = "PRED", y = "CWRES", xs = defx(), ys = defy(),
  xname = "value", ...)

wres_pred(df, x = "PRED", y = "WRES", xs = defx(), ys = defy(),
  xname = "value", ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{xname}{used to form x-axis label}

\item{...}{passed to \code{\link{scatt}} and \code{\link{layer_hs}}}
}
\description{
Residuals versus predicted values
}
\details{
Since this function creates a scatter plot,
both the \code{x} and \code{y} columns must
be numeric.

The y axis name is always the name of the residual
(e.g. "Weighted residual").  Use the \code{xname} argument
to add specific name and or unit to the dependent variable
(see the example).
}
\examples{

df <- dplyr::filter(superset2(), EVID==0)

cwres_pred(df, xname="MyDrug (ng/mL)")


}
\seealso{
\code{\link{geom_3s}}
}
