% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq.R
\name{res_q}
\alias{res_q}
\alias{wres_q}
\alias{cwres_q}
\title{QQ plot for conditional residuals}
\usage{
res_q(df, x = "RES", xs = defx(), ys = defy(), abline = c(0, 1), ...)

wres_q(df, x = "WRES", ...)

cwres_q(df, x = "CWRES", ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{abline}{numeric vector with two elements to be passed to \code{geom_abline}}

\item{...}{not used}
}
\description{
QQ plot for conditional residuals
}
\examples{

df <- dplyr::filter(superset2(), EVID==0)

cwres_q(df)

}
