% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rot_at}
\alias{rot_at}
\title{Rotate axis tick marks in a list of plots}
\usage{
rot_at(x, at = names(x), re = NULL, axis = c("x", "y"), ...)
}
\arguments{
\item{x}{a named list of gg or patchwork objects.}

\item{at}{a character vector of list names to rotate.}

\item{re}{a regular expression for selecting names to be used for \code{at}.}

\item{axis}{which axis to rotate.}

\item{...}{additional arguments passed to \code{\link[=rot_x]{rot_x()}} or \code{\link[=rot_y]{rot_y()}}.}
}
\description{
Pass in a list of gg or patchwork objects and rotate tick marks on x or y
axes.
}
\details{
Note that all plots in the list need to be named.
}
\examples{
data <- pmplots_data_id()

co <- c("STUDYc", "CPc", "RF")
etas <- paste0("ETA", 1:3)

x <- eta_cat(data, x = co, y = etas)
names(x)

x <- rot_at(x, at = "ETA1vRF", angle = 35)
x$ETA1vRF

x <- rot_at(x, re = "RF", vertical = TRUE)
x$ETA2vRF
x$ETA3vRF

}
\seealso{
\code{\link[=rot_xy]{rot_xy()}}, \code{\link[=rot_x]{rot_x()}}, \code{\link[=rot_y]{rot_y()}}.
}
