% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rot_xy}
\alias{rot_xy}
\alias{rot_xy.gg}
\alias{rot_xy.patchwork}
\alias{rot_xy.list}
\title{Rotate the left a right axes}
\usage{
rot_xy(x, ...)

\method{rot_xy}{gg}(x, ...)

\method{rot_xy}{patchwork}(x, ...)

\method{rot_xy}{list}(x, ...)
}
\arguments{
\item{x}{a gg object, patchwork object, or list of those objects.}

\item{...}{arguments passed to \code{\link[=rot_at]{rot_at()}} and \code{\link[=rot_x]{rot_x()}} or \code{\link[=rot_y]{rot_y()}}.}
}
\description{
Rotate the left a right axes
}
