% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rotx}
\alias{rotx}
\alias{roty}
\title{Rotate axis text}
\usage{
rotx(angle = 30, hjust = 1)

roty(angle = 30, hjust = 1)
}
\arguments{
\item{angle}{passed to \code{ggplot::element_text}}

\item{hjust}{passed to \code{ggplot::element_text}}
}
\description{
Rotate axis text
}
