% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_break}
\alias{str_break}
\title{Add a single newline to break a string across two lines}
\usage{
str_break(x, width = Inf)
}
\arguments{
\item{x}{a string}

\item{width}{the number of characters for the first of
two lines.}
}
\description{
Add a single newline to break a string across two lines
}
\examples{
pmplots:::str_break("Population prediction CX122341 concentration", 10)

}
\keyword{internal}
