% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{[.pm_opts}
\alias{[.pm_opts}
\title{Extract multiple pmplots options}
\usage{
\method{[}{pm_opts}(x, i, ..., exact = TRUE)
}
\arguments{
\item{x}{the \code{pm_opts} object}

\item{i}{a character option name to extract}

\item{...}{other option names to extract}

\item{exact}{not used}
}
\description{
Extract multiple pmplots options
}
\examples{
pm["scatter.size", "time.unit"]

}
