% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_time.R
\name{wres_time}
\alias{wres_time}
\alias{wres_tafd}
\alias{wres_tad}
\title{Plot weighted residuals versus time}
\usage{
wres_time(df, x = pm_axis_time(), y = pm_axis_wres(), ...)

wres_tafd(df, x = pm_axis_tafd(), y = pm_axis_wres(), ...)

wres_tad(df, x = pm_axis_tad(), y = pm_axis_wres(), ...)
}
\arguments{
\item{df}{data set to plot}

\item{x}{x-axis data in \code{\link{col_label}} format}

\item{y}{y-axis data in \code{\link{col_label}} format}

\item{...}{passed to \code{\link{res_time}} and eventually to
\code{\link{y_time}}}
}
\description{
Plot weighted residuals versus time
}
\details{
See \code{\link{res_time}} for other important implementation details.
}
\examples{
df <- pmplots_data_obs()

wres_time(df)

}
\seealso{
\code{\link{res_time}}, \code{\link{cwres_time}},
\code{\link{npde_time}}
}
