% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatt.R
\name{y_time}
\alias{y_time}
\title{Plot continuous data versus time}
\usage{
y_time(df, x = "TIME", y, xname = "Time", xunit = "hr", yname = NULL,
  xs = defx(), ys = defy(), log = FALSE, xby = NULL, ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{xname}{used to form x-axis title}

\item{xunit}{used to form x-axis title}

\item{yname}{used to form y-axis title}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{log}{if \code{TRUE}, y-axis will be log-transformed}

\item{xby}{interval for breaks on x-axis (time)}

\item{...}{passed to \code{\link{scatt}}}
}
\description{
Plot continuous data versus time
}
\details{
This function is intented for internal use.

When the y-axis label needs a unit, include it in
the \code{yname} specification.

Since this function creates a scatter plot,
both the \code{x} and \code{y} columns must
be numeric.

The \code{xby} argument is provided for
convenience to set breaks for time scales.
}
