% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{boxwork}
\alias{boxwork}
\title{Make boxplots}
\usage{
boxwork(df, x, y, xs = defcx(), ys = defy(), fill = "white", alpha = 1,
  hline = NULL, title = NULL, shown = TRUE, ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{fill}{passed to \code{geom_boxplot}}

\item{alpha}{passed to \code{geom_boxplot}}

\item{hline}{used to draw horizontal reference line}

\item{title}{passed to \code{ggtitle}}

\item{shown}{if \code{TRUE} numbers in each box are shown below tick labels}

\item{...}{not used}
}
\description{
Make boxplots
}
