% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{boxwork}
\alias{boxwork}
\title{Make boxplots}
\usage{
boxwork(df, x, y, xs = defcx(), ys = defy(), fill = "white",
  alpha = 1, hline = NULL, title = NULL, shown = TRUE, ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{xs}{see \code{\link{defcx}}}

\item{ys}{see \code{\link{defy}}}

\item{fill}{passed to \code{geom_boxplot}}

\item{alpha}{passed to \code{geom_boxplot}}

\item{hline}{used to draw horizontal reference line}

\item{title}{passed to \code{ggtitle}}

\item{shown}{if \code{TRUE} provide a numeric summary of each
box (see details)}

\item{...}{not used}
}
\description{
Make boxplots
}
\details{
Since this function creates a boxplot,
the \code{x} column must be character, factor
or logical and \code{y} column must
be numeric.

If \code{shown} is \code{TRUE}, a numeric summary of each
box is included
below each box.  In the summary, \code{n} is the number of
non-NA observations in the \code{y} column for that box and
\code{N} is the number of unique \code{ID} values for
that box.  An error will be generated if \code{ID} does
not exist in the plotting data frame when \code{shown} is
\code{TRUE}.  When \code{N} is equal to \code{n} in the
summary, only \code{n} is shown.
}
