% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_label}
\alias{col_label}
\title{Column // axis-label specification}
\usage{
col_label(x)
}
\arguments{
\item{x}{string encoding data column and axis title}
}
\description{
Column // axis-label specification
}
\examples{
col_label("CL // Clearance (L)")
col_label("CL//Clearance (L)")
col_label("CL $$ Clearance (L)")
col_label("CL !! Clearance (L)")
col_label("CL @@ Clearance (L)")

try(col_label("CL / Clearance (L)"))

col_label("CL")


}
