% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{cont_cat}
\alias{cont_cat}
\title{Plot continuous variable against a categorical variable}
\usage{
cont_cat(df, x, y, xs = defcx(), ys = defy(), ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character col//title for x-axis data; see \code{\link{col_label}}}

\item{y}{character col//title for y-axis data; see \code{\link{col_label}}}

\item{xs}{see \code{\link{defcx}}}

\item{ys}{see \code{\link{defy}}}

\item{...}{other arguments passed to \code{\link{boxwork}}}
}
\description{
Plot continuous variable against a categorical variable
}
\details{
Since this function creates a boxplot,
the \code{x} column must be character, factor
or logical and \code{y} column must
be numeric.

Summary numbers located below each box are described in
\code{\link{boxwork}}.
}
\examples{
df <- dplyr::filter(pmplots_data(), EVID==0)
df <- dplyr::distinct(df, ID, .keep_all = TRUE)

cont_cat(df, x="STUDYc//Study name", y="WT//Weight (kg)")

}
\seealso{
\code{\link{boxwork}}
}
