% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont.R
\name{cont_cont}
\alias{cont_cont}
\alias{cont_cont_list}
\title{Plot continuous variable versus continuous variable}
\usage{
cont_cont(df, x, y, xs = defx(), ys = defy(), ...)

cont_cont_list(df, x, y, ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character col//title for x-axis data; see \code{\link{col_label}}}

\item{y}{character col//title for y-axis data; see \code{\link{col_label}}}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{...}{passed to \code{\link{scatt}}  and \code{\link{layer_hs}}}
}
\description{
This function is primarily called by other functions.
\code{cont_cont_list} is a vectorized form of \code{cont_cont}.
}
\details{
Since this function creates a scatter plot,
both the \code{x} and \code{y} columns must
be numeric.

The \code{eta_cont} function can make a plot of ETA versus
continuous variable for multiple ETAs when a vector
of values for \code{y} are passed in.  Accordingly,
\code{eta_cont} always returns a list of plots of the same
length as \code{y}.
}
\examples{
df <- dplyr::filter(pmplots_data(), EVID==0)
df <- dplyr::distinct(df, ID, .keep_all = TRUE)

cont_cont(df, x="WT//Weight (kg)", y="HT//Height (cm)")

}
\seealso{
\code{\link{scatt}}
}
