% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{cont_hist}
\alias{cont_hist}
\alias{cont_hist_list}
\alias{pm_histogram}
\title{Generate a histogram plot}
\usage{
cont_hist(df, x, xs = defx(), fill = "black", col = "white",
  alpha = 0.6, y = "..count..", add_density = y == "..density..", ...)

cont_hist_list(df, x, ...)

pm_histogram(..., col = "white", fill = "black", alpha = 0.6)
}
\arguments{
\item{df}{the data frame containing plotting data}

\item{x}{the x column for \code{geom_histogram}}

\item{xs}{a list of information for the x axis}

\item{fill}{a character value passed to \code{geom_histogram}}

\item{col}{a character value passed to \code{geom_histogram}}

\item{alpha}{a numeric value passed to \code{geom_histogram}}

\item{y}{what to use for the y-axis on the histogram; can be
\code{"..count.."} or \code{"..density.."}}

\item{add_density}{if \code{TRUE}, a normal density line will
be plotted over the histogram using \code{\link{add_density}}}

\item{...}{passed to \code{geom_histogram} and \code{add_density}}
}
\description{
\code{cont_hist_list} is a vectorized version
of \code{cont_hist}.  \code{pm_histogram} is a generic histogram
function that is called by other functions in pmplots.
}
\examples{

df <- pmplots_data_id()

cont_hist(df, x = "WT//Weight (kg)")


}
