% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{cont_hist}
\alias{cont_hist}
\alias{cont_hist_list}
\alias{pm_histogram}
\title{Generate a histogram plot}
\usage{
cont_hist(
  df,
  x,
  xs = defx(),
  y = "count",
  add_density = y == "density",
  add_layers = TRUE,
  ...
)

cont_hist_list(df, x, ...)

pm_histogram(
  ...,
  col = opts$histogram.col,
  fill = opts$histogram.fill,
  alpha = opts$histogram.alpha
)
}
\arguments{
\item{df}{the data frame containing plotting data.}

\item{x}{the x column for \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}.}

\item{xs}{a list of information for the x axis.}

\item{y}{what to use for the y-axis on the histogram; can be
\code{"count"} or \code{"density"}.}

\item{add_density}{if \code{TRUE}, a normal density line will
be plotted over the histogram using \code{\link[=add_density]{add_density()}}.}

\item{add_layers}{extra layers will be added only if \code{TRUE}.}

\item{...}{passed to \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}} and \code{\link[=add_density]{add_density()}}.}

\item{col}{a character value passed to \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}.}

\item{fill}{a character value passed to \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}.}

\item{alpha}{a numeric value passed to \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}.}
}
\value{
A single plot.
}
\description{
\code{cont_hist_list} is a vectorized version
of \code{cont_hist}.  \code{pm_histogram} is a generic histogram
function that is called by other functions in pmplots.
}
\examples{

data <- data.frame(WT = rnorm(1000,80,20))

cont_hist(data, x = "WT//Weight (kg)")

}
