% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefab.R
\name{cwres_covariate}
\alias{cwres_covariate}
\alias{cwres_covariate_list}
\title{Plot CWRES versus covariates}
\usage{
cwres_covariate(df, x, ncol = 2, tag_levels = NULL)

cwres_covariate_list(df, x)
}
\arguments{
\item{df}{a data frame to plot.}

\item{x}{character \verb{col//title} for covariates to plot on x-axis;
see \link{col_label}.}

\item{ncol}{passed to \code{\link[=pm_grid]{pm_grid()}}.}

\item{tag_levels}{passed to \code{\link[patchwork:plot_annotation]{patchwork::plot_annotation()}}.}
}
\description{
Plot CWRES versus covariates
}
\examples{
data <- pmplots_data_id()
cont <- c("WT//Weight (kg)", "ALB//Albumin (mg/dL)")
cat <- c("RF//Renal function", "CPc//Child-Pugh")
cwres_covariate(data, x = c(cont, cat), tag_levels = "A")

}
\seealso{
\code{\link[=npde_covariate]{npde_covariate()}}, \code{\link[=eta_covariate]{eta_covariate()}}
}
