% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefab.R
\name{cwres_panel}
\alias{cwres_panel}
\alias{cwres_panel_list}
\title{Plot a panel of CWRES diagnostic plots}
\usage{
cwres_panel(data, time_unit = "days", xname = "value", tag = FALSE)

cwres_panel_list(data, time_unit = "days", xname = "value")
}
\description{
Plot a panel of CWRES diagnostic plots
}
