% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displays.R
\name{cwres_panel}
\alias{cwres_panel}
\alias{cwres_panel_list}
\title{Create a display of CWRES diagnostic plots}
\usage{
cwres_panel(
  df,
  xname = "value",
  unit_time = "hours",
  unit_tad = "hours",
  xby_time = NULL,
  xby_tad = NULL,
  tag_levels = NULL
)

cwres_panel_list(
  df,
  xname = "value",
  unit_time = "hours",
  unit_tad = "hours",
  xby_time = NULL,
  xby_tad = NULL
)
}
\arguments{
\item{df}{a data frame to plot.}

\item{xname}{passed to \code{\link[=npde_pred]{npde_pred()}}.}

\item{unit_time}{passed to \code{\link[=npde_tad]{npde_tad()}} as \code{xunit}.}

\item{unit_tad}{passed to \code{\link[=npde_time]{npde_time()}} as \code{xunit}.}

\item{xby_time}{passed to \code{\link[=npde_time]{npde_time()}} as \code{xby}.}

\item{xby_tad}{passed to \code{\link[=npde_tad]{npde_tad()}} as \code{xby}.}

\item{tag_levels}{passed to \code{\link[patchwork:plot_annotation]{patchwork::plot_annotation()}}.}
}
\value{
\code{cwres_panel()} returns a single graphic with the following panels:
\itemize{
\item \code{CWRES} versus \code{TIME} via \code{\link[=cwres_time]{cwres_time()}}
\item \code{CWRES} versus \code{TAD} via \code{\link[=cwres_tad]{cwres_tad()}}
\item \code{CWRES} versus \code{PRED} via \code{\link[=cwres_pred]{cwres_pred()}}
\item \code{CWRES} histogram via \code{\link[=cwres_hist]{cwres_hist()}}
\item \code{CWRES} quantile-quantile plot via \code{\link[=cwres_q]{cwres_q()}}
}

\code{cwres_panel_list()} returns a list of the individual plots that are
incorporated into the \code{cwres_panel()} output. Each element of the list
is named for the plot in that position: \code{time}, \code{tad}, \code{pred}, \code{hist}
\code{q}. See \strong{Examples} for how you can work with that list.
}
\description{
Get a single graphic of basic \code{CWRES} diagnostics (\code{cwres_panel()}) or get
the component plots in a list that can be arranged by the user
(\code{cwres_panel_list()}) . See \code{\link[=cwres_covariate]{cwres_covariate()}} for plotting \code{CWRES}
versus covariates.
}
\examples{
data <- pmplots_data_obs()
cwres_panel(data, tag_levels = "A")

l <- cwres_panel_list(data)
with(l, (q+hist) / pred)

}
\seealso{
\code{\link[=npde_panel]{npde_panel()}}, \code{\link[=npde_covariate]{npde_covariate()}}
}
