% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{defcx}
\alias{defcx}
\title{Default setting for discrete x-axis scale}
\usage{
defcx(...)
}
\arguments{
\item{...}{arguments for \code{scale_x_discrete}}
}
\description{
A named list of the formal arguments for \code{scale_x_discrete}.   This
function helps you to create a fully populated list.  For any
function that uses this as a default, you can also
create your own named list for arguments you want to update using
this default as a base.
}
\details{
In the named list, the name is the argument name and the value
is the argument value.
}
\examples{
defcx()

}
