% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{defy}
\alias{defy}
\title{Default setting for continuous y-axis scale}
\usage{
defy(..., trans = deprecated())
}
\arguments{
\item{...}{arguments for \code{scale_y_continuous}}

\item{trans}{deprecated; use \code{transform} argument instead.}
}
\description{
A named list of the formal arguments for \code{scale_y_continuous}.  This
function helps you to create a fully populated list.  For any
function that uses this as a default, you can also
create your own named list for arguments you want to update, using
this default as a base.
}
\examples{
defy(transform="log")

}
