% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_pred.R
\name{dv_pred}
\alias{dv_pred}
\alias{dv_ipred}
\title{Plot DV versus predicted values}
\usage{
dv_pred(df, x = "PRED", y = "DV", xname = yname, yname = "value",
  xs = defx(), ys = defy(), loglog = FALSE, prefix = "Population", ...)

dv_ipred(df, x = "IPRED", ..., prefix = "Individual")
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{xname}{used to form x-axis label}

\item{yname}{used to form y-axis label}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{loglog}{if \code{TRUE}, x- and y-axes will be log-transformed}

\item{prefix}{used internally}

\item{...}{passed to \code{\link{scatt}} and \code{\link{layer_as}}}
}
\description{
Plot DV versus predicted values
}
\details{
Since this function creates a scatter plot,
both the \code{x} and \code{y} columns must
be numeric.
}
\examples{
df <- dplyr::filter(superset2(), EVID==0)

dv_pred(df)

dv_ipred(df, yname="MyDrug (ng/mL)")

}
