% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatt.R
\name{dv_pred}
\alias{dv_pred}
\alias{dv_ipred}
\title{Plot DV versus predicted values}
\usage{
dv_pred(df, x = "DV", y = "PRED", xname = "value", yname = xname,
  xs = defx(), ys = defy(), loglog = FALSE, prefix = "Population", ...)

dv_ipred(df, y = "IPRED", ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{xname}{used to form x-axis label}

\item{yname}{used to form y-axis label}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{loglog}{if \code{TRUE}, x- and y-axes will be log-transformed}

\item{prefix}{used internally}

\item{...}{passed to \code{\link{scatt}}}
}
\description{
Plot DV versus predicted values
}
\examples{
df <- dplyr::filter(superset2(), EVID==0)
dv_pred(df)

}
