% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_pred.R
\name{dv_pred}
\alias{dv_pred}
\alias{dv_ipred}
\alias{dv_preds}
\title{Plot DV versus predicted values}
\usage{
dv_pred(
  df,
  x = pm_axis_pred(),
  y = pm_axis_dv(),
  yname = "value",
  xname = "value",
  ys = list(),
  xs = ys,
  loglog = FALSE,
  scales = c("fixed", "free", "null"),
  logbr = c("full", "half", "null"),
  ...
)

dv_ipred(df, x = pm_axis_ipred(), ...)

dv_preds(df, ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{yname}{used to form y-axis label}

\item{xname}{used to form x-axis label}

\item{ys}{see \code{\link{defy}}}

\item{xs}{see \code{\link{defx}}; note that \code{xs} defaults to \code{ys} so (by
default) the scale configuration will be identical; pass both \code{xs} and \code{ys}
to have them independently configured}

\item{loglog}{if \code{TRUE}, x- and y-axes will be log-transformed}

\item{scales}{if \code{TRUE}, then the x- and y- axes will be forced
to have the same limits}

\item{logbr}{when using log scale, should the tick marks be at \code{full}-log
intervals or \code{half}-log intervals? If you pass \code{null}, the default scales
will be used (which might be identical to \code{full}). Use \code{xs} and \code{ys} to
pass custom scales.}

\item{...}{passed to \code{\link{scatt}} and \code{\link{layer_as}}}
}
\value{
A single plot.
}
\description{
Plot DV versus predicted values
}
\details{
Since this function creates a scatter plot,
both the \code{x} and \code{y} columns must
be numeric.

\code{dv_preds} returns a list of two plots, with
the result of \code{dv_pred} in the first position
and the result of \code{dv_ipred} in the
second position.  In this case, \code{...} are
passed to both functions.
}
\examples{
df <- pmplots_data_obs()

dv_pred(df)

dv_ipred(df, yname="MyDrug (ng/mL)")

dv_preds(df, yname = "MyDrug (ng/mL)")

}
