% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_pred_ipred.R
\name{dv_pred_ipred}
\alias{dv_pred_ipred}
\alias{dv_pred_ipred_impl}
\alias{do_dv_pred_ipred}
\title{Create plots showing DV, PRED, and IPRED by individual}
\usage{
dv_pred_ipred(
  data,
  id_per_plot = 9,
  id_col = "ID",
  nrow = NULL,
  ncol = NULL,
  fun = NULL,
  ...
)

dv_pred_ipred_impl(
  data,
  x = pm_axis_time(),
  dv = pm_axis_dv(),
  pred = pm_axis_pred(),
  ipred = pm_axis_ipred(),
  id_col = "USUBJID",
  xbreaks = waiver(),
  xunit = opts$time.unit,
  xlab = NULL,
  ylab = NULL,
  angle = NULL,
  font_size = 7,
  margin = 1,
  legend.position = "top",
  pred_lty = 2,
  ipred_lty = 1,
  lwd = 0.7,
  size = opts$scatter.size,
  dv_shape = 1,
  dv_line = FALSE,
  dv_lwd = 0.5,
  scales = "free",
  log_y = FALSE,
  use_theme = theme_plain(),
  dv_color = "black",
  ipred_color = "firebrick",
  pred_color = "darkslateblue",
  ncol = NULL,
  nrow = NULL,
  axis.text.rel = NULL,
  fun = NULL
)

do_dv_pred_ipred(data, options = list())
}
\arguments{
\item{data}{the data frame to plot}

\item{id_per_plot}{number of subjects to include in each page}

\item{id_col}{the column name for the subject identifier; the value will
be displayed in the plot strip}

\item{nrow}{number of rows in the plot grid; passed to \link[ggplot2:facet_wrap]{ggplot2::facet_wrap}}

\item{ncol}{number of columns in the plot grid; passed to \link[ggplot2:facet_wrap]{ggplot2::facet_wrap}}

\item{fun}{a function accepting a gg object as argument and returning
an updated gg object; experimental}

\item{...}{additional arguments passed to \link{dv_pred_ipred_impl}}

\item{x}{the time-axis column, in \link{col_label} format; the title portion is
used for the x-axis title along with \code{xunit}; see also the \code{xlab} argument}

\item{dv}{the \code{DV} column, in \link{col_label} format; the title portion is used
for the y-axis title; see also the \code{ylab} argument}

\item{pred}{the name of the \code{PRED} column; \link{col_label} format is allowed, but
the label portion is discarded}

\item{ipred}{the name of the \code{IPRED} column; \link{col_label} format is allowed,
but the label portion is discarded}

\item{xbreaks}{x-axis breaks; passed to \link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous}}

\item{xunit}{used to form x-axis title only if \code{xlab} is not provided}

\item{xlab}{x-axis title; if not \code{NULL}, passed to \link[ggplot2:labs]{ggplot2::xlab}}

\item{ylab}{y-axis title; if not \code{NULL}, passed to \link[ggplot2:labs]{ggplot2::ylab}}

\item{angle}{rotation angle for x-axis tick labels; passed to \link{rot_x}}

\item{font_size}{the font size for strip text}

\item{margin}{the top and bottom margin for the plot strip; passed to
\link[ggplot2:element]{ggplot2::margin} as \code{t} and \code{b}}

\item{legend.position}{passed to \link[ggplot2:theme]{ggplot2::theme}}

\item{pred_lty}{\code{PRED} linetype; passed to \link[ggplot2:geom_path]{ggplot2::geom_line}}

\item{ipred_lty}{\code{IPRED} linetype; passed to \link[ggplot2:geom_path]{ggplot2::geom_line}}

\item{lwd}{line width for \code{PRED} and \code{IPRED}; passed to \link[ggplot2:geom_path]{ggplot2::geom_line}}

\item{size}{size of shape for \code{DV}; passed to \link[ggplot2:geom_point]{ggplot2::geom_point}}

\item{dv_shape}{shape for \code{DV}; passed to \link[ggplot2:geom_point]{ggplot2::geom_point}}

\item{dv_line}{logical; if \code{TRUE} then a line is added to the plot connecting
\code{DV} points}

\item{dv_lwd}{line width for \code{DV}; passed to \link[ggplot2:geom_path]{ggplot2::geom_line} as \code{lwd}}

\item{scales}{passed to \link[ggplot2:facet_wrap]{ggplot2::facet_wrap}}

\item{log_y}{logical; if \code{TRUE} then y-axis is shown in log-scale}

\item{use_theme}{a theme to use for the plot}

\item{dv_color}{color to use for \code{DV} points}

\item{ipred_color}{color to use for \code{IPRED} line}

\item{pred_color}{color to use for \code{PRED} line}

\item{axis.text.rel}{relative text size for axis text; use this to selectively
decrease font size for axis tick labels}

\item{options}{a named list of options to pass to \link{dv_pred_ipred}}
}
\description{
\code{DV} is plotted with a symbol while \code{PRED} and \code{IPRED} are plotted with lines.
The plot is faceted by unique individual identifier (like \code{ID} or \code{USUBJID}).
}
\details{
if both \code{nrow} and \code{ncol} are supplied and numeric, \code{id_per_plot} will be set
to \code{nrow*ncol}.
}
\examples{

data <- pmplots_data_obs()

p <- dv_pred_ipred(data, ylab="Concentration (ng/mL)", nrow=3, ncol=3)

}
