% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_time.R
\name{dv_time}
\alias{dv_time}
\alias{dv_tafd}
\alias{dv_tad}
\title{Plot DV versus time}
\usage{
dv_time(df, x = "TIME", y = "DV", xname = "Time", xunit = "hr",
  yname = NULL, group = "ID", xs = defx(), ys = defy(), log = FALSE,
  xby = NULL, ...)

dv_tafd(df, x = "TAFD", ..., xname = "Time after first dose")

dv_tad(df, x = "TAD", ..., xname = "Time after dose")
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{xname}{used to form x-axis label}

\item{xunit}{time units; used to form x-axis label}

\item{yname}{used to form y-axis label}

\item{group}{passed to \code{\link{scatt}}}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{log}{if \code{TRUE}, y-axis will be log-transformed}

\item{xby}{interval for x-axis breaks}

\item{...}{passed to \code{\link{scatt}}}
}
\description{
Plot DV versus time
}
\details{
Since this function creates a scatter plot,
both the \code{x} and \code{y} columns must
be numeric.

If the data set includes a \code{BLQ} column,
the values in the \code{y} column are
set to \code{NA} when \code{BLQ} is
not equal to \code{0}.
}
\examples{
df <- dplyr::filter(superset2(), EVID==0)

dv_time(df, yname="MyDrug (ng/mL)")

dv_time(df, yname="MyDrug (ng/mL)", xunit="day")

dv_time(df, log=TRUE, col="STUDYc")

# Add a rug
dv_time(df) + geom_rug(data = function(x) dplyr::filter(x, BLQ > 0))

}
