% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatt.R
\name{dv_time}
\alias{dv_time}
\title{Plot DV versus time}
\usage{
dv_time(df, x = "TIME", y = "DV", xunit = "hr", yname = NULL,
  group = "ID", xs = defx(), ys = defy(), log = FALSE, ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{xunit}{time units; used to form x-axis label}

\item{yname}{used to form y-axis label}

\item{group}{passed to \code{\link{scatt}}}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{log}{if \code{TRUE}, y-axis will be log-transformed}

\item{...}{passed to \code{\link{scatt}}}
}
\description{
Plot DV versus time
}
\examples{
df <- dplyr::filter(superset2(), EVID==0)
dv_time(df)
dv_time(df, log=TRUE, col="STUDYc")

}
