% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefab.R
\name{eta_covariate}
\alias{eta_covariate}
\alias{eta_covariate_list}
\title{Create ETA versus covariate displays}
\usage{
eta_covariate(
  df,
  x,
  y,
  ncol = 2,
  tag_levels = NULL,
  byrow = NULL,
  transpose = FALSE
)

eta_covariate_list(df, x, y, transpose = FALSE)
}
\arguments{
\item{df}{a data frame to plot.}

\item{x}{character \verb{col//title} for covariates to plot on x-axis;
see \link{col_label}.}

\item{y}{character \verb{col//title} for ETAs to plot on y-axis; see \link{col_label}.}

\item{ncol}{passed to \code{\link[=pm_grid]{pm_grid()}}.}

\item{tag_levels}{passed to \code{\link[patchwork:plot_annotation]{patchwork::plot_annotation()}}.}

\item{byrow}{passed through \code{\link[=pm_grid]{pm_grid()}}.}

\item{transpose}{logical; if \code{TRUE}, output will be transposed to group
plots by the covariate, rather than the \code{ETA}; see \strong{Examples}.}
}
\value{
\code{eta_covariate()} returns a list of plots arranged in graphics as a
\code{patchwork} object using \code{\link[=pm_grid]{pm_grid()}} and \code{eta_covariate_list()} the same
plots, but unarranged as a list of lists.

When \code{transpose} is \code{FALSE} (default), plots in a single graphic are grouped
by the \code{ETA} name (e.g., \code{ETA1}) the names of the list reflect that \code{ETA}
name. When \code{transpose} is \code{TRUE}, the graphics are grouped by column names
passed via \code{x} and the names of the list reflect those covariate data names.
See \strong{Examples}.
}
\description{
Get a single graphic of \code{ETA} versus continuous and / or categorical
covariates (\code{eta_covariate()}) or the component plots as a list
(\code{eta_covariate_list()}) that can be arranged by the user.
}
\details{
Pass \code{ncol = NULL} or another non-numeric value to bypass arranging plots
coming from \code{eta_covariate()}.
}
\examples{
data <- pmplots_data_id()
etas <- c("ETA1//ETA-CL", "ETA2//ETA-V")
cont <- c("WT//Weight (kg)", "ALB//Albumin (mg/dL)")
cat <- c("RF//Renal function", "CPc//Child-Pugh")
eta_covariate(data, x = c(cont, cat), y = etas, tag_levels = "A")

eta_covariate(data, cont, y = etas)
eta_covariate(data, cont, y = etas, transpose = TRUE)

eta_covariate_list(data, x = cat, y = etas)
eta_covariate_list(data, x = cat, y = etas, transpose = TRUE)

}
\seealso{
\code{\link[=npde_covariate]{npde_covariate()}}, \code{\link[=cwres_covariate]{cwres_covariate()}}
}
