% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefab.R
\name{eta_covariate}
\alias{eta_covariate}
\alias{eta_covariate_list}
\title{Plot ETAs versus covariates}
\usage{
eta_covariate(df, x, y, ncol = 2, tag_levels = NULL, byrow = NULL)

eta_covariate_list(df, x, y)
}
\arguments{
\item{df}{a data frame to plot.}

\item{x}{character \verb{col//title} for covariates to plot on x-axis;
see \link{col_label}.}

\item{y}{character \verb{col//title} for ETAs to plot on y-axis; see \link{col_label}.}

\item{ncol}{passed to \code{\link[=pm_grid]{pm_grid()}}.}

\item{tag_levels}{passed to \code{\link[patchwork:plot_annotation]{patchwork::plot_annotation()}}.}

\item{byrow}{passed through \code{\link[=pm_grid]{pm_grid()}}.}
}
\description{
Plot ETAs versus covariates
}
\examples{
data <- pmplots_data_id()
etas <- c("ETA1//ETA-CL", "ETA2//ETA-V")
cont <- c("WT//Weight (kg)", "ALB//Albumin (mg/dL)")
cat <- c("RF//Renal function", "CPc//Child-Pugh")
eta_covariate(data, x = c(cont, cat), y = etas, tag_levels = "A")


}
\seealso{
\code{\link[=npde_covariate]{npde_covariate()}}, \code{\link[=cwres_covariate]{cwres_covariate()}}
}
