% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_hist.R
\name{eta_hist}
\alias{eta_hist}
\title{ETA histograms}
\usage{
eta_hist(df, x, xs = defx(), add_density = FALSE, ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character col//label for x-axis data; see \code{\link{col_label}}}

\item{xs}{see \code{\link{defx}}}

\item{add_density}{if \code{TRUE}, a normal density line will
be plotted on top of the histogram via \code{\link{add_density}}}

\item{...}{other arguments for \code{geom_histogram}}
}
\value{
A single plot when a single value for \code{x}
and \code{y} are supplied; a list of plots of either \code{x}
or \code{y} have length greater than 1.
}
\description{
ETA histograms
}
\details{
The \code{x} column must be numeric.  This function
calls \code{\link{cont_hist}}.
}
\examples{

df <- pmplots_data_id()

etas <- c("ETA1//ETA-CL", "ETA2//ETA-V2", "ETA3//ETA-KA")

eta_hist(df, etas)

}
