% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_hist.R
\name{eta_hist}
\alias{eta_hist}
\title{ETA histograms}
\usage{
eta_hist(df, x, xs = defx(), fill = "black", col = "white", alpha = 0.6,
  add_density = FALSE, ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character col//label for x-axis data; see \code{\link{col_label}}}

\item{xs}{see \code{\link{defx}}}

\item{fill}{passed to \code{geom_histogram}}

\item{col}{passed to \code{geom_histogram}}

\item{alpha}{passed to \code{geom_histogram}}

\item{...}{other arguments for \code{geom_histogram}}
}
\description{
ETA histograms
}
\details{
The \code{x} column must be numeric.  This function
calls \code{\link{cont_hist}}.
}
\examples{

df <- dplyr::filter(pmplots_data(), EVID==0)
id <- dplyr::distinct(df, ID, .keep_all = TRUE)

etas <- c("ETA1//ETA-CL", "ETA2//ETA-V2", "ETA3//ETA-KA")

eta_hist(id, etas)

}
