% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{eta_hist}
\alias{eta_hist}
\title{ETA histograms}
\usage{
eta_hist(df, x, xs = defx(), fill = "black", col = "white", alpha = 0.6,
  ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character col//label for x-axis data; see \code{\link{col_label}}}

\item{xs}{see \code{\link{defx}}}

\item{fill}{passed to \code{geom_histogram}}

\item{col}{passed to \code{geom_histogram}}

\item{alpha}{passed to \code{geom_histogram}}

\item{...}{other arguments for \code{geom_histogram}}
}
\description{
ETA histograms
}
\examples{
df <- dplyr::filter(superset2(), EVID==0)
etas <- c("ETA1//ETA-CL", "ETA2//ETA-V2", "ETA3//ETA-KA")
eta_hist(df, etas)

}
