% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{layer}
\alias{layer}
\alias{layer_h}
\alias{layer_s}
\alias{layer_a}
\alias{layer_hs}
\alias{layer_sh}
\alias{layer_as}
\alias{layer_sa}
\alias{layer_dots}
\alias{gs}
\alias{ga}
\alias{gh}
\title{Layer functions}
\usage{
layer_h(x, hline = gh(), add_layers = TRUE, ...)

layer_s(x, smooth = gs(), add_layers = TRUE, ...)

layer_a(x, abline = ga(), add_layers = TRUE, ...)

layer_hs(x, ...)

layer_sh(x, ...)

layer_as(x, ...)

layer_sa(x, ...)

layer_dots(x, ...)

gs(method = "loess", se = FALSE, lty = 2, lwd = 1.35,
  col = .ggblue, ...)

ga(intercept = 0, slope = 1, lwd = 1.35, col = "darkgrey", ...)

gh(yintercept = 0, lwd = 1.35, col = "darkgrey", ...)
}
\arguments{
\item{x}{a \code{ggplot} object}

\item{hline}{list of arguments for \code{geom_hline}}

\item{add_layers}{if \code{FALSE} no layers are added from
\code{layer_s}, \code{layer_a}, \code{layer_h}, or
combinations}

\item{...}{passed to layering functions and geoms}

\item{smooth}{list of arguments for \code{geom_smooth}}

\item{abline}{list of arguments for \code{geom_abline}}

\item{method}{passed to the appropriate \code{geom_}}

\item{se}{passed to the appropriate \code{geom_}}

\item{lty}{passed to the appropriate \code{geom_}}

\item{lwd}{passed to the appropriate \code{geom_}}

\item{col}{passed to the appropriate \code{geom_}}

\item{intercept}{passed to \code{geom_abline}}

\item{slope}{passed to \code{geom_abline}}

\item{yintercept}{passed to \code{geom_hline}}
}
\description{
Layer functions
}
\details{
Function names can be decoded as: \code{h} indicates horizontal
reference line, \code{s} indicates smoothing line,
\code{a} indicates \code{abline} (typically a line of
identity). The order of the the codes indicates the
order in which the layers are applied. For example,
\code{layer_hs} means to first add a horizontal reference
line and then add a smoothing line.  Likewise, \code{layer_s}
adds a smoother, \code{layer_a} adds identity line, and
\code{layer_y} adds a horizontal reference line.

\code{gs}, \code{ga}, and \code{gh} are helper functions to create
default arguments to \code{geom_smooth}, \code{geom_abline}, and
\code{geom_hline}, respectively.  The \code{gx} series
of functions are not exported.
}
\examples{
df <- dplyr::filter(pmplots_data(), EVID==0 & BLQ==0)

cwresi_time(df) + geom_3s()

pmplots:::gs()


}
\seealso{
\code{\link{pm_smooth}}, \code{\link{pm_hline}},
\code{\link{pm_abline}}
}
