% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefab.R
\name{npde_covariate}
\alias{npde_covariate}
\alias{npde_covariate_list}
\title{Plot NPDE versus covariates}
\usage{
npde_covariate(df, x, ncol = 2, tag_levels = NULL, byrow = NULL)

npde_covariate_list(df, x)
}
\arguments{
\item{df}{a data frame to plot.}

\item{x}{character \verb{col//title} for covariates to plot on x-axis;
see \link{col_label}.}

\item{ncol}{passed to \code{\link[=pm_grid]{pm_grid()}}.}

\item{tag_levels}{passed to \code{\link[patchwork:plot_annotation]{patchwork::plot_annotation()}}.}

\item{byrow}{passed through \code{\link[=pm_grid]{pm_grid()}}.}
}
\description{
Plot NPDE versus covariates
}
\examples{
data <- pmplots_data_id()
cont <- c("WT//Weight (kg)", "ALB//Albumin (mg/dL)")
cat <- c("RF//Renal function", "CPc//Child-Pugh")
npde_covariate(data, x = c(cont, cat), y = etas, tag_levels = "A")

}
\seealso{
\code{\link[=cwres_covariate]{cwres_covariate()}}, \code{\link[=eta_covariate]{eta_covariate()}}
}
