% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displays.R
\name{npde_panel}
\alias{npde_panel}
\alias{npde_panel_list}
\title{Create a display of NPDE diagnostic plots}
\usage{
npde_panel(
  df,
  xname = "value",
  unit_time = "hours",
  unit_tad = "hours",
  xby_time = NULL,
  xby_tad = NULL,
  tag_levels = NULL
)

npde_panel_list(
  df,
  xname = "value",
  unit_time = "hours",
  unit_tad = "hours",
  xby_time = NULL,
  xby_tad = NULL
)
}
\arguments{
\item{df}{a data frame to plot.}

\item{xname}{passed through \code{\link[=npde_pred]{npde_pred()}}.}

\item{unit_time}{passed through \code{\link[=npde_tad]{npde_tad()}} as \code{xunit}.}

\item{unit_tad}{passed through \code{\link[=npde_time]{npde_time()}} as \code{xunit}.}

\item{xby_time}{passed to \code{\link[=npde_time]{npde_time()}} as \code{xby}.}

\item{xby_tad}{passed to \code{\link[=npde_tad]{npde_tad()}} as \code{xby}.}

\item{tag_levels}{passed to \code{\link[patchwork:plot_annotation]{patchwork::plot_annotation()}}.}
}
\value{
\code{npde_panel()} returns a single graphic as a \code{patchwork} object with the
following panels:
\itemize{
\item \code{NPDE} versus \code{TIME} via \code{\link[=npde_time]{npde_time()}}
\item \code{NPDE} versus \code{TAD} via \code{\link[=npde_tad]{npde_tad()}}
\item \code{NPDE} versus \code{PRED} via \code{\link[=npde_pred]{npde_pred()}}
\item \code{NPDE} histogram via \code{\link[=npde_hist]{npde_hist()}}
\item \code{NPDE} quantile-quantile plot via \code{\link[=npde_q]{npde_q()}}
}

\code{npde_panel_list()} returns a list of the individual plots that are
incorporated into the \code{npde_panel()} output. Each element of the list
is named for the plot in that position: \code{time}, \code{tad}, \code{pred}, \code{hist}
\code{q}. See \strong{Examples} for how you can work with that list.
}
\description{
Get a single graphic of basic \code{NPDE} diagnostics (\code{npde_panel()}) or get the
component plots in a list (\code{npde_panel_list()}) that can be arranged by the
user. See \code{\link[=npde_covariate]{npde_covariate()}} for plotting \code{NPDE} versus covariates.
}
\examples{
data <- pmplots_data_obs()
npde_panel(data, tag_levels = "A")

l <- npde_panel_list(data)
names(l)
with(l, (q+hist) / pred, tag_levels = "a")

}
\seealso{
\code{\link[=cwres_panel]{cwres_panel()}}
}
