% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_time.R
\name{npde_time}
\alias{npde_time}
\alias{npde_tad}
\alias{npde_tafd}
\title{Plot NPDE versus time}
\usage{
npde_time(df, x = pm_axis_time(), y = pm_axis_npde(), ...,
  hline = npde_ref())

npde_tad(df, x = pm_axis_tad(), y = pm_axis_npde(), ...)

npde_tafd(df, x = pm_axis_tafd(), y = pm_axis_npde(), ...)
}
\arguments{
\item{df}{data set to plot}

\item{x}{x-axis data in \code{\link{col_label}} format}

\item{y}{y-axis data in \code{\link{col_label}} format}

\item{...}{passed to \code{\link{res_time}} and eventually to
\code{\link{y_time}}}

\item{hline}{a list of arguments to pass to \code{geom_hline} specifying
aesthetics to use}
}
\description{
Plot NPDE versus time
}
\details{
See \code{\link{res_time}} for other important implementation details.
}
\examples{
df <- pmplots_data_obs()

npde_time(df)

}
\seealso{
\code{\link{res_time}}, \code{\link{cwres_time}},
\code{\link{wres_time}}
}
