% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_pairs.R
\name{pairs_plot}
\alias{pairs_plot}
\alias{eta_pairs}
\title{Pairs plots using ggpairs}
\usage{
pairs_plot(x, etas, bins = 15, alpha = 0.6, fill = "black",
  col = "grey", label_fun = label_parse_label, upper_fun = NULL,
  lower_fun = NULL, ...)

eta_pairs(...)
}
\arguments{
\item{x}{plotting data.frame}

\item{etas}{character \code{col//label} for pairs data; see \link{col_label}}

\item{bins}{passed to \link[ggplot2:geom_histogram]{ggplot2::geom_histogram}}

\item{alpha}{passed to \link[ggplot2:geom_histogram]{ggplot2::geom_histogram}}

\item{fill}{passed to \link[ggplot2:geom_histogram]{ggplot2::geom_histogram}}

\item{col}{passed to \link[ggplot2:geom_histogram]{ggplot2::geom_histogram}}

\item{label_fun}{labeler function that gets passed to \link[GGally:ggpairs]{GGally::ggpairs};
the default is based on \link{parse_label} and thus allows latex
expressions in the label (see examples)}

\item{upper_fun}{function to use for \code{upper} argument}

\item{lower_fun}{function to use for \code{lower} argument}

\item{...}{passed to \link[GGally:ggpairs]{GGally::ggpairs}}
}
\value{
The result from a \code{ggpairs} call (a single plot).
}
\description{
This funciton is a wrapper to \link[GGally:ggpairs]{GGally::ggpairs} with customized
functions for upper and lower off-diagonal panels.
}
\details{
This function requires the \code{GGally} package to be installed.

When the length of \code{etas} is one, arguments are passed to \link{eta_hist} and
that result is returned.
}
\examples{

id <- pmplots_data_id()

etas <- c("ETA1//ETA-CL", "ETA2//ETA-VC", "ETA3//ETA-KA")

eta_pairs(id,etas)

df <- data.frame(x = rnorm(1000), y = rnorm(1000))

pairs_plot(df, c("x", "y"))

df2 <- dplyr::tibble(x = rnorm(100), y = x^2)

pairs_plot(df2, c("x//x", "y//x$^2$"))

}
