% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_axis_data.R
\docType{data}
\name{pm_axis}
\alias{pm_axis}
\alias{pm_axis_data}
\title{Get defaults for plot axes}
\format{An object of class \code{list} of length 3.}
\usage{
pm_axis(what)

pm_axis_data
}
\arguments{
\item{what}{the axis identifier}
}
\description{
Get defaults for plot axes
}
\examples{
pm_axis("cwres")

}
\keyword{datasets}
