% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_axis_data.R
\name{pm_axis_functions}
\alias{pm_axis_functions}
\alias{pm_axis_tad}
\alias{pm_axis_time}
\alias{pm_axis_tafd}
\alias{pm_axis_res}
\alias{pm_axis_wres}
\alias{pm_axis_cwres}
\alias{pm_axis_cwresi}
\alias{pm_axis_npde}
\alias{pm_axis_pred}
\alias{pm_axis_ipred}
\alias{pm_axis_dv}
\alias{pm_col_tad}
\alias{pm_col_id}
\title{Functions to generate axis data and other standard column names.}
\usage{
pm_axis_tad(xunit = NULL)

pm_axis_time(xunit = NULL)

pm_axis_tafd(xunit = NULL)

pm_axis_res()

pm_axis_wres()

pm_axis_cwres()

pm_axis_cwresi()

pm_axis_npde()

pm_axis_pred(xname = NULL)

pm_axis_ipred(xname = NULL)

pm_axis_dv(yname = NULL)

pm_col_tad()

pm_col_id()
}
\arguments{
\item{xunit}{an optional time unit to be glued into the title.}

\item{xname}{an optional name to be glued into the title.}

\item{yname}{an optional name to be glued into the title.}
}
\description{
Functions to generate axis data and other standard column names.
}
\details{
The \verb{pm_axis_*} functions call \code{pmplots:::mk_col_title()}. When time
units (\code{xunit}) are passed, they will be wrapped in parens. The
\verb{pm_col_*} functions return standard column names or their alias if
set (see \code{\link[=pm_aliases]{pm_aliases()}}).

\code{ID} is occasionally used inside pmplots functions as a subject
identifier. \code{pm_col_id()} is a wrapper around
\code{getOption("mrg.id_col", "ID")}; this option can be to globally
override \code{ID} as the default subject identifier.
}
\examples{
pm_axis_time()
pm_axis_time("h")
pm_axis_pred("concentration")
pm_col_id()

}
\seealso{
\code{\link[=pm_axis]{pm_axis()}}
}
