% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_axis_data.R
\name{pm_axis_functions}
\alias{pm_axis_functions}
\alias{pm_axis_tad}
\alias{pm_axis_time}
\alias{pm_axis_tafd}
\alias{pm_axis_res}
\alias{pm_axis_wres}
\alias{pm_axis_cwres}
\alias{pm_axis_cwresi}
\alias{pm_axis_npde}
\alias{pm_axis_pred}
\alias{pm_axis_ipred}
\alias{pm_axis_dv}
\title{Functions to generate axis data}
\usage{
pm_axis_tad(xunit = NULL)

pm_axis_time(xunit = NULL)

pm_axis_tafd(xunit = NULL)

pm_axis_res()

pm_axis_wres()

pm_axis_cwres()

pm_axis_cwresi()

pm_axis_npde()

pm_axis_pred(xname = NULL)

pm_axis_ipred(xname = NULL)

pm_axis_dv(yname = NULL)
}
\arguments{
\item{xunit}{an optional time unit to be glued into the title.}

\item{xname}{an optional name to be glued into the title.}

\item{yname}{an optional name to be glued into the title.}
}
\description{
Functions to generate axis data
}
\details{
These functions call  either \code{pmplots:::mk_col_title()} or
\code{pmplots:::mk_col_title_opt()}. When time units (\code{xunit}) are
passed, they will be wrapped in parens.
}
\examples{
pm_axis_time()
pm_axis_time("h")
pm_axis_pred("concentration")

}
\seealso{
\code{\link[=pm_axis]{pm_axis()}}
}
