% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_axis_data.R
\name{pm_axis_functions}
\alias{pm_axis_functions}
\alias{pm_axis_tad}
\alias{pm_axis_time}
\alias{pm_axis_tafd}
\alias{pm_axis_res}
\alias{pm_axis_wres}
\alias{pm_axis_cwres}
\alias{pm_axis_cwresi}
\alias{pm_axis_npde}
\alias{pm_axis_pred}
\alias{pm_axis_ipred}
\alias{pm_axis_dv}
\title{Functions to generate axis data}
\usage{
pm_axis_tad()

pm_axis_time()

pm_axis_tafd()

pm_axis_res()

pm_axis_wres()

pm_axis_cwres()

pm_axis_cwresi()

pm_axis_npde()

pm_axis_pred()

pm_axis_ipred()

pm_axis_dv()
}
\description{
Functions to generate axis data
}
\details{
These functions call \code{pmplots:::mk_col_title}
}
\examples{
pmplots:::pm_axis_time()

}
