% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_browser.R
\name{pm_browser}
\alias{pm_browser}
\alias{pm_browser_show}
\title{pmplots browser object}
\usage{
pm_browser(quiet = FALSE, title = TRUE)

pm_browser_show(x, plot_name = NULL)
}
\arguments{
\item{quiet}{if \code{TRUE} then a summary of available plots
will be printed}

\item{title}{if \code{TRUE}, plotting code will be added as the plot title}

\item{x}{output from \code{pm_browser}}

\item{plot_name}{the name of a plot to show}
}
\description{
pmplots browser object
}
\seealso{
\code{\link{pm_gadget}}
}
