% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_browser.R
\name{pm_browser}
\alias{pm_browser}
\alias{pm_browser_show}
\title{pmplots browser}
\usage{
pm_browser(quiet = FALSE)

pm_browser_show(x, plot_name = NULL)
}
\arguments{
\item{quiet}{if `TRUE` then a summary of available plots
will be printed}

\item{x}{output from `pm_browser`}

\item{plot_name}{the name of a plot to show}
}
\description{
pmplots browser
}
