% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pm_grid}
\alias{pm_grid}
\title{Arrange a list of plots in a grid}
\usage{
pm_grid(x, ..., ncol = 2, tag_levels = NULL)
}
\arguments{
\item{x}{a list of plots.}

\item{...}{passed to \code{\link[patchwork:wrap_plots]{patchwork::wrap_plots()}}.}

\item{ncol}{passed to \code{\link[patchwork:wrap_plots]{patchwork::wrap_plots()}}.}

\item{tag_levels}{passed to \code{\link[patchwork:plot_annotation]{patchwork::plot_annotation()}}.}
}
\description{
This is a light wrapper around \code{\link[patchwork:wrap_plots]{patchwork::wrap_plots()}} to
arrange the plots and \code{\link[patchwork:plot_annotation]{patchwork::plot_annotation()}} to
optionally tag levels in the grid of plots.
}
\details{
The patchwork package must be installed to use this function.
}
\examples{
data <- pmplots_data_obs()

plot <- wres_cont(data, x = c("WT", "ALB"))

pm_grid(plot)

pm_grid(plot, tag_levels = "a")

}
