% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pm_grid}
\alias{pm_grid}
\title{Arrange a list of plots in a grid}
\usage{
pm_grid(x, ..., ncol = 2)
}
\arguments{
\item{x}{A list of plots.}

\item{...}{Passed to \code{\link[patchwork:wrap_plots]{patchwork::wrap_plots()}}.}

\item{ncol}{Passed to \code{\link[patchwork:wrap_plots]{patchwork::wrap_plots()}}.}
}
\description{
This is a ligth wrapper around \code{\link[patchwork:wrap_plots]{patchwork::wrap_plots()}}.
}
\details{
The patchwork package must be installed to use this function.
}
\examples{
data <- pmplots_data_obs()

plot <- wres_cont(data, x = c("WT", "ALB"))

pm_grid(plot)

}
