% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_plot.R
\name{pm_grid}
\alias{pm_grid}
\title{Arrange a list of plots in a grid}
\usage{
pm_grid(x, ...)
}
\arguments{
\item{x}{a list of plots}

\item{...}{passed to \code{\link[cowplot]{plot_grid}}}
}
\description{
Arrange a list of plots in a grid
}
\details{
The cowplot package must be installed to use this function.
}
