% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmp-gg-labs.R
\name{pm_label_columns}
\alias{pm_label_columns}
\title{Add axis label data to a data frame}
\usage{
pm_label_columns(df, spec, labs = list(), short_max = Inf)
}
\arguments{
\item{df}{a data frame to label.}

\item{spec}{a named list of label data; names correspond to columns
in the data used to make the plot; may also be a \code{yspec} object, which
will be converted to a named list through \code{yspec::ys_get_short_unit()}.}

\item{labs}{another object like \code{spec} containing of label data to override
names found in \code{spec}.}

\item{short_max}{passed to \code{\link[yspec:ys_get_short_unit]{yspec::ys_get_short_unit()}} when \code{spec} is a
\code{yspec} object.}
}
\value{
The data frame \code{df} with \code{pmp.axis.label} attributes set on the
labeled columns. If no columns in \code{spec} or \code{labs} match column names in
\code{df}, a warning is issued and \code{df} is returned unchanged.
}
\description{
This function adds candidate axis titles as an attribute on columns in
\code{data}. This attribute is intended to be specifically used for pmplot axis
labels, not to be confused with the label added by \code{yspec::ys_add_labels()}.
}
\seealso{
\code{\link[=pm_label_rm]{pm_label_rm()}}
}
