% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmp-gg-labs.R
\name{pm_label_rm}
\alias{pm_label_rm}
\title{Remove pmplot-specific axis label information}
\usage{
pm_label_rm(df)
}
\arguments{
\item{df}{a data frame to de-label.}
}
\value{
The data frame \code{df} with all \code{pmp.axis.label} attributes removed.
}
\description{
Remove pmplot-specific axis label information
}
\seealso{
\code{\link[=pm_label_columns]{pm_label_columns()}}
}
