% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\docType{data}
\name{pm_opts}
\alias{pm_opts}
\alias{pm_options}
\alias{pm}
\title{Set global plot options}
\format{
An object of class \code{pm_opts} (inherits from \code{environment}) of length 50.

An object of class \code{pm_opts} (inherits from \code{environment}) of length 50.
}
\usage{
pm_options(
  smooth.lwd = 1.35,
  smooth.col = .ggblue,
  smooth.lty = 2,
  smooth.method = "loess",
  scatter.size = 1.5,
  scatter.col = "black",
  scatter.alpha = 1,
  density.lwd = 1.35,
  density.col = .ggblue,
  density.lty = 2,
  hline.lwd = 1.35,
  hline.col = "darkgrey",
  hline.lty = 1,
  abline.lwd = 1.35,
  abline.col = "darkgrey",
  abline.lty = 1,
  histogram.fill = "black",
  histogram.alpha = 0.6,
  histogram.col = "white",
  boxplot.fill = "white",
  boxplot.alpha = 1,
  boxplot.hline.lwd = 1,
  boxplot.hline.lty = 2,
  boxplot.hline.col = "black",
  boxplot.outlier.shape = 19,
  qq.col = .ggblue,
  qq.alpha = 1,
  qq.size = 1.35,
  pairs.cor.size = 3,
  pairs.cor.prefix = "corr\\n",
  pairs.cor.col = .ggblue,
  pairs.cor.fontface = "bold",
  pairs.cor.digits = 2,
  pairs.cor.shown = TRUE,
  axis.title.short = FALSE,
  glue.xunit = TRUE,
  time.unit = "h",
  time.label = "Time",
  time.label.short = time.label,
  tafd.label = "Time after first dose",
  tafd.label.short = tafd.label,
  tad.label = "Time after dose",
  tad.label.short = tad.label
)

pm_opts

pm
}
\arguments{
\item{smooth.lwd}{line width for smoother}

\item{smooth.col}{line color for smoother}

\item{smooth.lty}{line type for smoother}

\item{smooth.method}{method to use for creating smoother}

\item{scatter.size}{point size for scatter plot}

\item{scatter.col}{character name of color for scatter plot}

\item{scatter.alpha}{transparency for points in scatter plot}

\item{density.lwd}{line width for density plot on histogram}

\item{density.col}{line color for density plot on histogram}

\item{density.lty}{line type for density plot on histogram}

\item{hline.lwd}{line width for horizontal reference line; see \code{\link[=layer_h]{layer_h()}}
and \code{\link[=gh]{gh()}}}

\item{hline.col}{line color for horizontal reference line; see \code{\link[=layer_h]{layer_h()}}
and \code{\link[=gh]{gh()}}}

\item{hline.lty}{line type for horizontal reference line; see \code{\link[=layer_h]{layer_h()}}
and \code{\link[=gh]{gh()}}}

\item{abline.lwd}{line width for diagonal reference line; see \code{\link[=layer_a]{layer_a()}}
and \code{\link[=ga]{ga()}}}

\item{abline.col}{line color for diagonal reference line; see \code{\link[=layer_a]{layer_a()}}
and \code{\link[=ga]{ga()}}}

\item{abline.lty}{line type for diagonal reference line; see \code{\link[=layer_a]{layer_a()}}
and \code{\link[=ga]{ga()}}}

\item{histogram.fill}{fill color for histograms}

\item{histogram.alpha}{alpha value for histograms}

\item{histogram.col}{line color for histograms}

\item{boxplot.fill}{fill color for histograms}

\item{boxplot.alpha}{alpha value for boxplots}

\item{boxplot.hline.lwd}{line width for horizontal reference line}

\item{boxplot.hline.lty}{line type for horizontal reference line}

\item{boxplot.hline.col}{line color for horizontal reference line}

\item{boxplot.outlier.shape}{shape for outliers in boxplots}

\item{qq.col}{point color for qq plots}

\item{qq.alpha}{alpha value for qq plots}

\item{qq.size}{point size for qq plots}

\item{pairs.cor.size}{font size for correlation text in upper panels of pairs
plots}

\item{pairs.cor.prefix}{prefix for correlation text in upper panels of pairs
plots}

\item{pairs.cor.col}{color for correlation text in upper panels of pairs plots}

\item{pairs.cor.fontface}{font face for correlation text in upper panels of
pairs plots}

\item{pairs.cor.digits}{number of digits to show when reporting correlation
in upper panels of pairs plots}

\item{pairs.cor.shown}{if \code{TRUE}, then report the number of non-missing
observations used to calculate correlation in upper panels of pairs plots}

\item{axis.title.short}{shorten standard axis titles}

\item{glue.xunit}{if \code{TRUE}, pmplots will attempt to glue \code{{xunit}} into
x-axis labels}

\item{time.unit}{default time unit}

\item{time.label}{default time label}

\item{time.label.short}{default short time label}

\item{tafd.label}{default tafd label}

\item{tafd.label.short}{default short tafd label}

\item{tad.label}{default tad label}

\item{tad.label.short}{default tafd label}
}
\description{
Set global plot options
}
\details{
\code{\link[=pm]{pm()}} and \code{\link[=pm_opts]{pm_opts()}} both refer to the same environment.

Global options can
be set in the environment with \code{pm$set(name = value)}.  There is also
a \code{.list} argument to \code{pm$set} that allows you to pass in a named list of
options to set (e.g. \code{pm$set(.list = list(smooth.col="red4"))}).

Values can be extracted with \code{pm$get("name")}.

Because it is an environment, the \code{$} operator can also be used to get and
set values (see examples).

Other methods in the environment include \code{pm$as.list()}, \code{pm$mget()},
\code{pm$reset()}.  \code{pm$self} refers to the environment itself. A list of
default settings can be obtained with \code{pm$defaults}.  Methods for the
\code{pm_opts} object include: \link{print.pm_opts}, \link{as.list.pm_opts},
\code{\link{$<-.pm_opts}}.
}
\examples{

pm$set(smooth.lwd = 2)

pm$smooth.lwd

pm$smooth.lwd <- 1.3

pm$smooth.lwd

pm$get("smooth.lwd")

pm$reset()

x <- pm$as.list()

\dontrun{
 defs <- pm$defaults
 defs$smooth.col <- "firebrick"
 pm$set(.list = defs)
}

}
\keyword{datasets}
