% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\docType{data}
\name{pm_opts}
\alias{pm_opts}
\alias{pm_options}
\title{Set plot options}
\format{An object of class \code{environment} of length 18.}
\usage{
pm_options(
  line.col = "#3366FF",
  smooth.lwd = 1.35,
  smooth.col = line.col,
  smooth.lty = 2,
  scatter.size = 1.5,
  density.col = line.col,
  density.lwd = 1.35,
  denstiy.lty = 2,
  hline.lwd = 1.35,
  hline.col = "darkgrey",
  abline.lwd = 1.35,
  abline.col = "darkgrey"
)

pm_opts
}
\arguments{
\item{smooth.lwd}{line width for smoother}

\item{smooth.col}{line color for smoother}

\item{smooth.lty}{line type for smoother}

\item{scatter.size}{point size for scatter plot}
}
\description{
Set plot options
}
\keyword{datasets}
