% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm-gg-labs.R
\name{pm_relabel}
\alias{pm_relabel}
\alias{pm_relabel.gg}
\alias{pm_relabel.list}
\title{Relabel a pmplots plot using a yspec object or named list}
\usage{
pm_relabel(x, ...)

\method{pm_relabel}{gg}(x, spec, labs = list(), ...)

\method{pm_relabel}{list}(x, spec, labs = list(), ...)
}
\arguments{
\item{x}{a gg object created through a \code{pmplots} function.}

\item{...}{additional arguments passed to \code{\link[=pm_gg_labs]{pm_gg_labs()}}.}

\item{spec}{a named list of label data; names correspond to columns
in the data used to make the plot; may also be a \code{yspec} object, which
will be converted to a named list through \code{\link[yspec:ys_get_short_unit]{yspec::ys_get_short_unit()}}.}

\item{labs}{another named list of label data to override names found in
\code{spec}.}
}
\description{
This function relabels \code{x} and \code{y} axis titles based on the data columns
used to create the plot, looking up in a named list or \code{yspec} object. This
is only for pmplot outputs; consider \code{\link[yspec:ys_gg_labs]{yspec::ys_gg_labs()}} for labeling
aesthetics in an arbitrary \code{gg} object.
}
\examples{
data <- pmplots_data_obs()

p <- dv_pred(data)
p

spec <- list(DV = "CX1123 concentration (ng/mL)")

p <- pm_relabel(p, spec)
p

}
\seealso{
\code{\link[=pm_gg_labs]{pm_gg_labs()}}
}
