% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm-gg-labs.R
\name{pm_relabel}
\alias{pm_relabel}
\alias{pm_relabel.gg}
\alias{pm_relabel.list}
\title{Relabel a plot using a yspec object or named list}
\usage{
pm_relabel(x, ...)

\method{pm_relabel}{gg}(x, spec, labs = list(), ...)

\method{pm_relabel}{list}(x, spec, labs = list(), ...)
}
\arguments{
\item{x}{a ggplot object or a list of ggplot objects.}

\item{...}{additional arguments passed to \code{\link[=pm_gg_labs]{pm_gg_labs()}}.}

\item{spec}{a named list of label data; names correspond to columns
in the data used to make the plot; may also be a \code{yspec} object, which
will be converted to a named list through \code{\link[yspec:ys_get_short_unit]{yspec::ys_get_short_unit()}}.}

\item{labs}{another named list of label data to override names found in
\code{spec}.}
}
\description{
Applies \code{\link[=pm_gg_labs]{pm_gg_labs()}} to a ggplot object or a list of ggplot objects,
updating axis and aesthetic labels by looking up column names in a named
list or \code{yspec} object.
}
\examples{
data <- pmplots_data_obs()

spec <- list(
  DV = "CX1123 concentration (ng/mL)", 
  PRED = "Population prediction (ng/mL)"
)

p <- dv_pred(data)

pm_relabel(p, spec)

}
\seealso{
\code{\link[=pm_gg_labs]{pm_gg_labs()}}, \code{\link[=pmp_gg_labs]{pmp_gg_labs()}}
}
