% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmp-gg-labs.R
\name{pmp_relabel_wrap}
\alias{pmp_relabel_wrap}
\title{Relabel facet strips in a wrapped pmplot}
\usage{
pmp_relabel_wrap(p, spec, labs = list(), short_max = Inf)
}
\arguments{
\item{p}{a ggplot object created by a \verb{wrap_*} pmplots function.}

\item{spec}{a named list of label data; names correspond to columns
in the data used to make the plot; may also be a \code{yspec} object, which
will be converted to a named list through \code{\link[yspec:ys_get_short_unit]{yspec::ys_get_short_unit()}}.}

\item{labs}{another named list of label data to override names found in
\code{spec}.}

\item{short_max}{passed to \code{\link[yspec:ys_get_short_unit]{yspec::ys_get_short_unit()}} when \code{spec} is a
\code{yspec} object.}
}
\value{
The plot \code{p} with updated facet strip labels.
}
\description{
This function relabels the facet strip labels in a plot created by a
\verb{wrap_*} function (e.g., \code{\link[=wrap_eta_cont]{wrap_eta_cont()}}, \code{\link[=wrap_cont_cont]{wrap_cont_cont()}}) by looking
up the facet variable names in a named list or \code{yspec} object. Variable
names are discovered automatically from the plot data. Names absent from
\code{spec} and \code{labs} are left unchanged.
}
\examples{
data <- pmplots_data_obs()

spec <- list(WT = "Weight (kg)", ALB = "Albumin (mg/dL)")

p <- wrap_eta_cont(data, x = c("WT", "ALB"), y = "ETA1//ETA1", scales = "free_x")

pmp_relabel_wrap(p, spec)

}
\seealso{
\code{\link[=pmp_relabel]{pmp_relabel()}}, \code{\link[=pmp_gg_labs]{pmp_gg_labs()}}
}
