% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aaaa.R
\name{pmplots}
\alias{pmplots}
\title{Plots for pharmacometrics}
\description{
Plots for pharmacometrics
}
\section{Look and feel}{


See the \link{pmtheme} help topic for functions you
can use to replicate the look and feel of plots generated
by this package.
}

\section{Function listing}{

\itemize{
\item Residuals versus time
\itemize{
\item \link{res_time}, \link{res_tad}, \link{res_tafd}
\item \link{wres_time}, \link{wres_tad}, \link{wres_tad}
\item \link{cwres_time}, \link{cwres_tad}, \link{cwres_tafd}
}
\item Residuals versus PRED
\itemize{
\item \link{res_pred}, \link{wres_pred}, \link{cwres_pred}
}
\item Residuals versus covariates
\itemize{
\item \link{res_cont}, \link{wres_cont}, \link{cwres_cont}
\item \link{res_cat}, \link{wres_cat}, \link{cwres_cat}
}
\item NPDE plots
\itemize{
\item \link{npde_time}, \link{npde_tad}, \link{npde_tafd}
\item \link{npde_pred}
\item \link{npde_hist}
\item \link{npde_q}
\item \link{npde_cont}
\item \link{npde_cat}
}
}
}

