% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmp-gg-labs.R
\name{relabel_at}
\alias{relabel_at}
\title{Relabel plots in a list using pm_relabel}
\usage{
relabel_at(x, at = names(x), spec = list(), labs = list(), re = NULL, ...)
}
\arguments{
\item{x}{a named list of gg objects.}

\item{at}{a character vector of list names to relabel.}

\item{spec}{a named list of label data; names correspond to columns in the
data used to make the plots; may also be a \code{yspec} object.}

\item{labs}{another object like \code{spec} containing of label data to override
names found in \code{spec}.}

\item{re}{a regular expression for selecting names to be used for \code{at}.}

\item{...}{additional arguments passed to \code{\link[=pm_relabel]{pm_relabel()}}.}
}
\description{
Pass in a named list of gg objects and apply \code{\link[=pm_relabel]{pm_relabel()}}.
}
\details{
Note that all plots in the list need to be named. When \code{re} is provided it
takes precedence over \code{at}.
}
\examples{
data <- pmplots_data_obs()

spec <- list(DV = "Observed (ng/mL)", PRED = "Population predicted (ng/mL)")

x <- list(p1 = dv_pred(data), p2 = dv_ipred(data))

x <- relabel_at(x, at = "p1", spec = spec)
x$p1

x <- relabel_at(x, re = "p", spec = spec)
x$p2

}
\seealso{
\code{\link[=pm_relabel]{pm_relabel()}}, \code{\link[=rot_at]{rot_at()}}.
}
