% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_cat.R
\name{res_cat}
\alias{res_cat}
\alias{wres_cat}
\alias{cwres_cat}
\alias{cwresi_cat}
\title{Plot residuals versus categorical variable}
\usage{
res_cat(df, x, y = pm_axis_res(), hline = 0, ...)

wres_cat(df, x, y = pm_axis_wres(), hline = 0, ...)

cwres_cat(df, x, y = pm_axis_cwres(), hline = 0, ...)

cwresi_cat(df, x, y = pm_axis_cwresi(), ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character col//title for x-axis data; see \code{\link{col_label}}}

\item{y}{character col//title for y-axis data; see \code{\link{col_label}}}

\item{hline}{where to draw horizontal refrence line}

\item{...}{passed to \code{\link{cont_cat}}}
}
\description{
These are vectorized functions.  See details.
}
\details{
These are vectorized functions; if either x or y
are a character vector (in col//title format), a list
of plots is returned.  If both x and y are length 1,
then a single plot object (not a list) is returned.

Since this function creates a boxplot,
the \code{x} column must be character, factor
or logical and \code{y} column must
be numeric.

Summary numbers located below each box are described in
\code{\link{boxwork}}.
}
\examples{

df <- dplyr::filter(pmplots_data(), EVID==0)

cwresi_cat(df, x="STUDYc//Study name")

}
\seealso{
\code{\link{boxwork}}, \code{\link{cont_cat}}
}
