% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_cont.R
\name{res_cont}
\alias{res_cont}
\alias{wres_cont}
\alias{cwres_cont}
\alias{cwresi_cont}
\alias{npde_cont}
\alias{npd_cont}
\title{Plot residuals versus continuous variables}
\usage{
res_cont(df, x, y = pm_axis_res(), xs = defx(), ys = defy(), ...)

wres_cont(df, x, y = pm_axis_wres(), xs = defx(), ys = defy(), ...)

cwres_cont(df, x, y = pm_axis_cwres(), xs = defx(), ys = defy(), ...)

cwresi_cont(df, x, y = pm_axis_cwresi(), ...)

npde_cont(df, x, y = pm_axis_npde(), ...)

npd_cont(df, x, y = pm_axis_npd(), ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character col//title for x-axis data; see
\code{\link{col_label}}}

\item{y}{character col//title for y-axis data; see
\code{\link{col_label}}}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{...}{passed to \code{\link{scatt}}  and
\code{\link{layer_hs}}}
}
\value{
A single plot when a single value for \code{x}
and \code{y} are supplied; a list of plots of either \code{x}
or \code{y} have length greater than 1.
}
\description{
These are vectorized functions; see details;
}
\details{
These are vectorized functions; if either x or y
are a character vector (in col//title format), a list
of plots is returned.  If both x and y are length 1,
then a single plot object (not a list) is returned.
}
\examples{
data <- pmplots_data_obs()

wres_cont(data, x = "CRCL//Creatinine clearance")

}
