% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_hist.R
\name{res_hist}
\alias{res_hist}
\alias{wres_hist}
\alias{cwres_hist}
\alias{cwresi_hist}
\alias{npde_hist}
\title{Histograms of residuals or NPDE}
\usage{
res_hist(df, ..., x = pm_axis_res(), xs = defx())

wres_hist(df, ..., x = pm_axis_wres(), y = "density")

cwres_hist(df, ..., x = pm_axis_cwres())

cwresi_hist(df, x = pm_axis_cwresi(), ...)

npde_hist(df, ..., x = pm_axis("npde"))
}
\arguments{
\item{df}{data frame to plot}

\item{...}{passed to \code{\link{cont_hist}}}

\item{x}{character name for x-axis data}

\item{xs}{see \code{\link{defx}}}

\item{y}{what to use for the y-axis on the histogram; can be
\code{"count"} or \code{"density"}}
}
\value{
A single plot.
}
\description{
Histograms of residuals or NPDE
}
\examples{
df <- pmplots_data_obs()

cwres_hist(df)

}
