% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatt.R
\name{res_pred}
\alias{res_pred}
\alias{cwres_pred}
\alias{wres_pred}
\title{Residuals versus predicted values}
\usage{
res_pred(df, x = "PRED", y = "RES", xs = defx(), ys = defy(),
  yname = "value", ...)

cwres_pred(df, x = "PRED", y = "CWRES", xs = defx(), ys = defy(),
  yname = "value", ...)

wres_pred(df, x = "PRED", y = "WRES", xs = defx(), ys = defy(),
  yname = "value", ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character naem for y-axis data}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{yname}{used to form y-axis label}

\item{...}{passed to \code{\link{scatt}}}
}
\description{
Residuals versus predicted values
}
\examples{
df <- dplyr::filter(superset2(), EVID==0)
cwres_pred(df)

}
