% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_pred.R
\name{res_pred}
\alias{res_pred}
\alias{wres_pred}
\alias{cwres_pred}
\alias{cwresi_pred}
\alias{npde_pred}
\title{Residuals or NPDE versus predicted values}
\usage{
res_pred(
  df,
  x = pm_axis_pred(),
  y = pm_axis_res(),
  xname = "value",
  xs = defx(),
  ys = defy(),
  ...
)

wres_pred(df, ..., y = pm_axis_wres())

cwres_pred(df, ..., y = pm_axis_cwres())

cwresi_pred(df, y = pm_axis_cwresi(), ...)

npde_pred(df, ..., y = pm_axis_npde(), hline = npde_ref())
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{xname}{glued into x-axis title}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{...}{passed to \code{\link{scatt}} and \code{\link{layer_hs}}}

\item{hline}{a list of parameters to pass to \code{geom_hline} specifying
where to locate a horizontal reference line aesthetics to use}
}
\value{
A single plot.
}
\description{
Residuals or NPDE versus predicted values
}
\details{
Since this function creates a scatter plot,
both the \code{x} and \code{y} columns must
be numeric.

The y axis name is always the name of the residual
(e.g. "Weighted residual").  Use the \code{xname} argument
to add specific name and or unit to the dependent variable
(see the example).

A loess smooth and a horizontal reference line are
layered on the plot.
}
\examples{

df <- pmplots_data_obs()

cwresi_pred(df, xname="MyDrug (ng/mL)")

}
\seealso{
\code{\link{geom_3s}}
}
