% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_pred.R
\name{res_pred}
\alias{res_pred}
\alias{wres_pred}
\alias{cwres_pred}
\alias{cwresi_pred}
\alias{npde_pred}
\title{Residuals or NPDE versus predicted values}
\usage{
res_pred(df, x = "PRED", y = "RES", xname = "value", yname = "Residual",
  xs = defx(), ys = defy(), ...)

wres_pred(df, ..., y = "WRES", yname = "Weighted residual")

cwres_pred(df, ..., y = "CWRES", yname = "Conditional weighted residual")

cwresi_pred(df, y = "CWRESI", ...)

npde_pred(df, ..., y = "NPDE",
  yname = "Normalized prediction distribution error")
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{xname}{used to form x-axis label}

\item{yname}{used to form y-axis label}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{...}{passed to \code{\link{scatt}} and \code{\link{layer_hs}}}
}
\description{
Residuals or NPDE versus predicted values
}
\details{
Since this function creates a scatter plot,
both the \code{x} and \code{y} columns must
be numeric.

The y axis name is always the name of the residual
(e.g. "Weighted residual").  Use the \code{xname} argument
to add specific name and or unit to the dependent variable
(see the example).
}
\examples{

df <- dplyr::filter(pmplots_data(), EVID==0)

cwres_pred(df, xname="MyDrug (ng/mL)")


}
\seealso{
\code{\link{geom_3s}}
}
