% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq.R
\name{res_q}
\alias{res_q}
\alias{wres_q}
\alias{cwres_q}
\alias{cwresi_q}
\alias{npde_q}
\title{QQ plot for conditional residuals or NPDE}
\usage{
wres_q(df, x = "WRES", xs = defx(), ys = defy(), abline = NULL, ...)

cwres_q(df, x = "CWRES", ...)

cwresi_q(df, x = "CWRESI", ...)

npde_q(df, x = "NPDE", ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{abline}{numeric vector with two elements to be passed to
\code{geom_abline}}

\item{...}{arguments passed back to \code{wres_q}}
}
\description{
QQ plot for conditional residuals or NPDE
}
\details{
The \code{x} column must be numeric.
}
\examples{

df <- dplyr::filter(pmplots_data_obs(), EVID==0, !is.na(IPRED))

cwresi_q(df)

}
