% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_time.R
\name{res_time}
\alias{res_time}
\alias{res_tafd}
\alias{res_tad}
\alias{wres_time}
\alias{wres_tafd}
\alias{wres_tad}
\alias{cwres_time}
\alias{cwresi_time}
\alias{cwres_tafd}
\alias{cwresi_tafd}
\alias{cwres_tad}
\alias{cwresi_tad}
\alias{npde_time}
\alias{npde_tad}
\alias{npde_tafd}
\title{Plot residuals or NPDE versus time}
\usage{
res_time(df, x = pm_axis_time(), y = pm_axis_res(), ...)

res_tafd(df, x = pm_axis_tafd(), ...)

res_tad(df, x = pm_axis_tad(), ...)

wres_time(df, x = pm_axis_time(), y = pm_axis_wres(), ...)

wres_tafd(df, x = pm_axis_tafd(), ...)

wres_tad(df, x = pm_axis_tad(), ...)

cwres_time(df, x = pm_axis_time(), y = pm_axis_cwres(), ...)

cwresi_time(df, y = pm_axis_cwresi(), ...)

cwres_tafd(df, x = pm_axis_tafd(), ...)

cwresi_tafd(df, y = pm_axis_cwresi(), ...)

cwres_tad(df, x = pm_axis_tad(), ...)

cwresi_tad(df, y = pm_axis_tad(), ...)

npde_time(df, y = pm_axis_npde(), ..., hline = npde_ref())

npde_tad(df, x = pm_axis_tad(), ...)

npde_tafd(df, x = pm_axis_tafd(), ...)
}
\arguments{
\item{df}{data set to plot}

\item{x}{character name of x-axis data}

\item{y}{character name of y-axis data}

\item{...}{passed to \code{\link{y_time}} and
\code{\link{layer_hs}}}

\item{hline}{a list of arguments to pass to \code{geom_hline} specifying
aesthetics to use}
}
\description{
Plot residuals or NPDE versus time
}
\details{
Functions are provided for plotting \code{RES}, \code{WRES},
and \code{CWRES} versus \code{TIME}, \code{TAFD}, and
\code{TAD}.  Plots are generated using \code{\link{y_time}},
which then calls \code{\link{scatt}}.

By default, the time unit is assumed
to be hours (\code{hr}).  See the \code{xunit} argument
to \code{\link{y_time}} to change the time unit.

See the \code{xby} argument to \code{\link{y_time}} for a
convenient way to change the breaks for the x-axis (time).

Since this function creates a scatter plot,
both the \code{x} and \code{y} columns must
be numeric.
}
\examples{
df <- dplyr::filter(pmplots_data(), EVID==0)

cwresi_time(df)

cwresi_time(df, yname = "CWRES")

cwresi_time(df, xunit="day")

wres_time(df, xby=48)

wres_time(df) + geom_3s()

}
\seealso{
\code{\link{y_time}}, \code{\link{geom_3s}}
}
