% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_time.R
\name{res_time}
\alias{res_time}
\alias{res_tafd}
\alias{res_tad}
\alias{wres_time}
\alias{wres_tafd}
\alias{wres_tad}
\alias{cwres_time}
\alias{cwresi_time}
\alias{cwres_tafd}
\alias{cwresi_tafd}
\alias{cwres_tad}
\alias{cwresi_tad}
\alias{npde_time}
\alias{npde_tad}
\alias{npde_tafd}
\title{Plot residuals or NPDE versus time}
\usage{
res_time(df, x = "TIME", y = "RES", xname = "Time", yname = "Residual",
  ...)

res_tafd(df, x = "TAFD", ..., xname = "Time after first dose")

res_tad(df, x = "TAD", ..., xname = "Time after dose")

wres_time(df, x = "TIME", y = "WRES", yname = "Weighted residual",
  xname = "Time", ...)

wres_tafd(df, x = "TAFD", ..., xname = "Time after first dose")

wres_tad(df, x = "TAD", ..., xname = "Time after dose")

cwres_time(df, x = "TIME", y = "CWRES",
  yname = "Conditional weighted residual", xname = "Time", ...)

cwresi_time(df, y = "CWRESI", ...)

cwres_tafd(df, x = "TAFD", ..., xname = "Time after first dose")

cwresi_tafd(df, y = "CWRESI", ...)

cwres_tad(df, x = "TAD", ..., xname = "Time after dose")

cwresi_tad(df, y = "CWRESI", ...)

npde_time(df, y = "NPDE", ...,
  yname = "Normalized prediction distribution error", hline = NULL)

npde_tad(df, x = "TAD", ..., xname = "Time after dose")

npde_tafd(df, x = "TAFD", ..., xname = "Time after first dose")
}
\arguments{
\item{df}{data set to plot}

\item{x}{character name of x-axis data}

\item{y}{character name of y-axis data}

\item{xname}{used for x-axis label}

\item{yname}{used for y-axis label}

\item{...}{passed to \code{\link{y_time}} and
\code{\link{layer_hs}}}
}
\description{
Plot residuals or NPDE versus time
}
\details{
Functions are provided for plotting \code{RES}, \code{WRES},
and \code{CWRES} versus \code{TIME}, \code{TAFD}, and
\code{TAD}.  Plots are generated using \code{\link{y_time}},
which then calls \code{\link{scatt}}.

By default, the time unit is assumed
to be hours (\code{hr}).  See the \code{xunit} argument
to \code{\link{y_time}} to change the time unit.

For all of these functions, \code{yname} is the
full specification of the y-axis title.

See the \code{xby} argument to \code{\link{y_time}} for a
convenient way to change the breaks for the x-axis (time).

Since this function creates a scatter plot,
both the \code{x} and \code{y} columns must
be numeric.
}
\examples{
df <- dplyr::filter(pmplots_data(), EVID==0)

cwres_time(df)

cwres_time(df, yname = "CWRES")

cwres_time(df, xunit="day")

wres_time(df, xby=48)

wres_time(df) + geom_3s()

}
\seealso{
\code{\link{y_time}}, \code{\link{geom_3s}}
}
