% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatt.R
\name{res_time}
\alias{res_time}
\alias{wres_time}
\alias{cwres_time}
\alias{cwres_tad}
\alias{wres_tad}
\alias{res_tad}
\title{Plot residuals versus time}
\usage{
res_time(df, what = "Residual", xwhat = "Time", x = "TIME", y = "RES",
  ...)

wres_time(df, what = "Weighted residual", xwhat = "Time", x = "TIME",
  y = "WRES", ...)

cwres_time(df, what = "Conditional weighted residual", x = "TIME",
  y = "CWRES", ...)

cwres_tad(df, what = "Conditional weighted residual",
  xwhat = "Time after dose", x = "TAD", y = "CWRES", ...)

wres_tad(df, what = "Weighted residual", xwhat = "Time after dose",
  x = "TAD", y = "WRES", ...)

res_tad(df, what = "Residual", xwhat = "Time after dose", x = "TAD",
  y = "RES", ...)
}
\arguments{
\item{df}{data set to plot}

\item{what}{used for y-axis label}

\item{xwhat}{used for x-axis label (not typically used)}

\item{x}{character name of x-axis data}

\item{y}{character name of y-axis data}

\item{...}{passed to \code{\link{y_time}}}
}
\description{
Plot residuals versus time
}
\examples{
df <- dplyr::filter(superset2(), EVID==0)
cwres_time(df)

}
\seealso{
\code{\link{y_time}}
}
