% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatt.R
\name{res_time}
\alias{res_time}
\alias{wres_time}
\alias{cwres_time}
\alias{cwres_tad}
\alias{wres_tad}
\alias{res_tad}
\title{Plot residuals versus time}
\usage{
res_time(df, yname = "Residual", x = "TIME", y = "RES", ...)

wres_time(df, yname = "Weighted residual", x = "TIME", y = "WRES", ...)

cwres_time(df, yname = "Conditional weighted residual", x = "TIME",
  y = "CWRES", ...)

cwres_tad(df, yname = "Conditional weighted residual",
  xname = "Time after dose", x = "TAD", y = "CWRES", ...)

wres_tad(df, yname = "Weighted residual", xname = "Time after dose",
  x = "TAD", y = "WRES", ...)

res_tad(df, yname = "Residual", xname = "Time after dose", x = "TAD",
  y = "RES", ...)
}
\arguments{
\item{df}{data set to plot}

\item{yname}{used for y-axis label}

\item{x}{character name of x-axis data}

\item{y}{character name of y-axis data}

\item{...}{passed to \code{\link{y_time}}}

\item{xname}{used for x-axis label}
}
\description{
Plot residuals versus time
}
\details{
By default, the time unit is assumed
to be hours (\code{hr}).  See the \code{xunit} argument
to \code{\link{y_time}} to change the time unit.
}
\examples{
df <- dplyr::filter(superset2(), EVID==0)

cwres_time(df)

cwres_time(df, xunit="day")

}
\seealso{
\code{\link{y_time}}
}
