% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_time.R
\name{res_time}
\alias{res_time}
\alias{res_tafd}
\alias{res_tad}
\title{Plot residuals versus time}
\usage{
res_time(df, x = pm_axis_time(), y = pm_axis_res(), ...)

res_tafd(df, x = pm_axis_tafd(), y = pm_axis_res(), ...)

res_tad(df, x = pm_axis_tad(), y = pm_axis_res(), ...)
}
\arguments{
\item{df}{data set to plot}

\item{x}{x-axis data in \code{\link{col_label}} format}

\item{y}{y-axis data in \code{\link{col_label}} format}

\item{...}{passed to \code{\link{y_time}} and
\code{\link{layer_hs}}}
}
\description{
These plots have \code{RES} on the y axis and
some time-related column on the x-axis, including \code{TIME},
\code{TAD}, or \code{TAFD}.
}
\details{
Plots are generated using \code{\link{y_time}},
which then calls \code{\link{scatt}}.

By default, the time unit is assumed
to be hours (\code{hr}).  See the \code{xunit} argument
to \code{\link{y_time}} to change the time unit.

See the \code{xby} argument to \code{\link{y_time}} for a
convenient way to change the breaks for the x-axis (time).

Since this function creates a scatter plot,
both the \code{x} and \code{y} columns must
be numeric.
}
\examples{
df <- dplyr::filter(pmplots_data(), EVID==0)

res_time(df)


}
\seealso{
\code{\link{y_time}}, \code{\link{geom_3s}}
}
