% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rot_x}
\alias{rot_x}
\alias{rot_y}
\title{Rotate axis text}
\usage{
rot_x(angle = 30, hjust = 1)

rot_y(angle = 30, hjust = 1)
}
\arguments{
\item{angle}{passed to \code{ggplot::element_text}}

\item{hjust}{passed to \code{ggplot::element_text}}
}
\description{
Rotate axis text
}
\examples{
data <- pmplots_data_obs()

dv_pred(data) + rot_x()

}
