% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatt.R
\name{scatt}
\alias{scatt}
\title{Scatter plot function.}
\usage{
scatt(df, x, y, xs = defx(), ys = defy(), smooth = TRUE,
  identity = FALSE, hline = NULL, title = NULL, group = NULL,
  col = NULL, scale_col = scale_color_brewer(palette = "Set2", name = ""),
  ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{smooth}{if \code{TRUE} a loess smooth will be drawn}

\item{identity}{if \code{TRUE} a line of identity will be drawn}

\item{hline}{if numeric, a horizontal reference line will be drawn}

\item{title}{character, passed to \code{ggtitle}}

\item{group}{character name of grouping variable}

\item{col}{character name of coloring variable}

\item{scale_col}{color scale}

\item{...}{not used}

\item{hline_col}{color if \code{hline} is drawn}
}
\description{
Scatter plot function.
}
\details{
Since this function creates a scatter plot,
both the \code{x} and \code{y} columns must
be numeric.
}
