% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatt.R
\name{scatt}
\alias{scatt}
\title{Scatter plot function}
\usage{
scatt(df, x, y, xs = defx(), ys = defy(), title = NULL, group = NULL,
  col = NULL, scale_col = scale_color_brewer(palette = "Set2", name = ""),
  ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{title}{character, passed to \code{ggtitle}}

\item{group}{character name of grouping variable}

\item{col}{not used}

\item{scale_col}{not used}

\item{...}{not used}
}
\description{
Scatter plot function
}
\details{
Since this function creates a scatter plot,
both the \code{x} and \code{y} columns must
be numeric.
}
