% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont.R
\name{scatt}
\alias{scatt}
\title{Scatter plot function}
\usage{
scatt(
  df,
  x,
  y,
  xs = defx(),
  ys = defy(),
  title = NULL,
  group = NULL,
  col = NULL,
  plot_id = FALSE,
  size = pm_opts$scatter.size,
  alpha = pm_opts$scatter.alpha,
  scale_col = scale_color_brewer(palette = "Set2", name = ""),
  ...
)
}
\arguments{
\item{df}{Data frame to plot.}

\item{x}{Character name for x-axis data.}

\item{y}{Character name for y-axis data.}

\item{xs}{See \code{\link[=defx]{defx()}}.}

\item{ys}{See \code{\link[=defy]{defy()}}.}

\item{title}{Character, passed to \code{\link[ggplot2:labs]{ggplot2::ggtitle()}}.}

\item{group}{Character name of grouping variable.}

\item{col}{Character name of variable to color the points.}

\item{plot_id}{If \code{TRUE} then subject IDs are plotted rather than points;
see the \code{size} argument - the size may need to be increased when plotting
IDs.}

\item{size}{Passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} or \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}.}

\item{alpha}{Passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{scale_col}{Discrete scale to use for coloring the points (see default).}

\item{...}{Not used.}
}
\value{
A single plot.
}
\description{
Scatter plot function
}
\details{
Since this function creates a scatter plot, both the \code{x} and \code{y} columns must
be numeric.
}
