% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont.R
\name{scatt}
\alias{scatt}
\title{Scatter plot function}
\usage{
scatt(
  df,
  x,
  y,
  xs = defx(),
  ys = defy(),
  title = NULL,
  group = NULL,
  plot_id = FALSE,
  size = pm_opts$scatter.size,
  col = pm_opts$scatter.col,
  mapping = aes(),
  ...
)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{xs}{see \code{\link[=defx]{defx()}}}

\item{ys}{see \code{\link[=defy]{defy()}}}

\item{title}{character, passed to \code{\link[ggplot2:labs]{ggplot2::ggtitle()}}}

\item{group}{character name of grouping variable}

\item{plot_id}{if \code{TRUE} then subject IDs are plotted rather than points;
see the \code{size} argument - the size may need to be increased when plotting IDs}

\item{size}{passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} or \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}}

\item{col}{character name of variable to color the points}

\item{mapping}{Additional list of mappings to be used for the plot; this
should be an object generated with \code{\link[ggplot2:aes]{ggplot2::aes()}}, \code{\link[ggplot2:aes_]{ggplot2::aes_string()}},
or \code{\link[ggplot2:aes_]{ggplot2::aes_()}}.}

\item{...}{not used}
}
\value{
A single plot.
}
\description{
Scatter plot function
}
\details{
Since this function creates a scatter plot,
both the \code{x} and \code{y} columns must
be numeric.
}
