% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitplot.R
\name{split_plot}
\alias{split_plot}
\title{Split and plot}
\usage{
split_plot(df, fun, sp = get_split_col(df), ...)
}
\arguments{
\item{df}{data frame to split and plot}

\item{fun}{function to use to plot}

\item{sp}{character name of column to split; if \code{df} is a
grouped data frame, leave this argument missing and the; regardless
of how this is specified, \code{sp} should name a single column
split column will be determined from the groups}

\item{...}{passed to \code{fun}}
}
\value{
A list of plots.
}
\description{
Split and plot
}
\examples{

df <- pmplots_data_obs()

require(dplyr)

df \%>\% group_by(STUDYc) \%>\% split_plot(dv_pred)

}
