% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitplot.R
\name{split_plot}
\alias{split_plot}
\title{Split and plot}
\usage{
split_plot(df, sp, fun = "cont_cont", ...)
}
\arguments{
\item{df}{data frame to split and plot}

\item{sp}{character name of column to split}

\item{fun}{function to use to plot}

\item{...}{passed to \code{fun}}
}
\description{
Split and plot
}
\examples{
df <- dplyr::filter(superset2(), EVID==0)
split_plot(df, sp="STUDYc", fun=dv_pred)

}
