% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_plots.R
\name{wrap_plots}
\alias{wrap_plots}
\alias{wrap_cont_cont}
\alias{wrap_cont_time}
\alias{wrap_res_time}
\alias{wrap_eta_cont}
\alias{wrap_hist}
\alias{wrap_dv_preds}
\title{Faceted plots}
\usage{
wrap_cont_cont(
  df,
  x,
  y,
  ...,
  fun = cont_cont,
  title = NULL,
  scales = "free_y",
  ncol = NULL,
  use_labels = FALSE,
  label_fun = label_parse_label
)

wrap_cont_time(df, ..., x = pm_axis_time())

wrap_res_time(df, ..., x = pm_axis_time())

wrap_eta_cont(df, x, y, scales = "fixed", ...)

wrap_hist(
  df,
  x,
  title = NULL,
  scales = "free_x",
  ncol = NULL,
  use_labels = FALSE,
  label_fun = label_parse_label,
  ...
)

wrap_dv_preds(
  df,
  ...,
  title = "Predicted {yname}",
  xname = "",
  scales = "fixed"
)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{x-axis data in \link{col_label} format}

\item{y}{y-axis data in \link{col_label} format}

\item{...}{passed to \code{fun}}

\item{fun}{the plotting function}

\item{title}{a title to use for the axis with faceting groups}

\item{scales}{passed to \code{facet_wrap}}

\item{ncol}{passed to \code{facet_wrap}}

\item{use_labels}{if \code{TRUE}, the label part of \code{col_label} will
be used in the strip; the column name is used otherwise}

\item{label_fun}{labeller function; passed to \link[ggplot2:facet_wrap]{ggplot2::facet_wrap}; the
default is based on \link{parse_label} and allows latex markup in the label}

\item{xname}{placeholder}
}
\description{
For these plots, data sets made long with respect to several
y-axis variables and then plotted and faceted with
\link[ggplot2:facet_wrap]{ggplot2::facet_wrap}.
}
\details{
\code{wrap_cont_cont} is a general function used by the others to
create faceted plots of two continuous variables.  \code{wrap_cont_time}
plots several continuous variables versus time.  \code{wrap_res_time}
plots several different residuals (or NPDE) versus time. \code{wrap_eta_cont}
plots etas versus a continuous covariate. \code{wrap_hist} creates a faceted
histogram plot.
}
