% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatt.R
\name{y_time}
\alias{y_time}
\title{Plot continuous data versus time.}
\usage{
y_time(df, x = "TIME", y, unit = "hours", xs = defx(), ys = defy(),
  what, xwhat = "Time", log = FALSE, ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{unit}{time unit, used to for x-axis label}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{what}{used to form y-axis title}

\item{xwhat}{used to form y-axis title}

\item{log}{if \code{TRUE}, y-axis will be log-transformed}

\item{...}{passed to \code{\link{scatt}}}
}
\description{
Plot continuous data versus time.
}
