% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatt.R
\name{y_time}
\alias{y_time}
\title{Plot continuous data versus time.}
\usage{
y_time(df, x = "TIME", y, xname = "Time", xunit = "hr", yname = NULL,
  xs = defx(), ys = defy(), log = FALSE, ...)
}
\arguments{
\item{df}{data frame to plot}

\item{x}{character name for x-axis data}

\item{y}{character name for y-axis data}

\item{xname}{used to form x-axis title}

\item{xunit}{used to form x-axis title}

\item{yname}{used to form y-axis title}

\item{xs}{see \code{\link{defx}}}

\item{ys}{see \code{\link{defy}}}

\item{log}{if \code{TRUE}, y-axis will be log-transformed}

\item{...}{passed to \code{\link{scatt}}}
}
\description{
Plot continuous data versus time.
}
